nrf52840_pac::cc_host_rgf

Module host_iot_kprtl_lock

This write-once register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.

Structs

HOST_IOT_KPRTL_LOCK_SPEC

This write-once register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.

R

Register HOST_IOT_KPRTL_LOCK reader

W

Register HOST_IOT_KPRTL_LOCK writer

Enums

HOST_IOT_KPRTL_LOCK_A

This register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.

Type Aliases

HOST_IOT_KPRTL_LOCK_R

Field HOST_IOT_KPRTL_LOCK reader - This register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.

HOST_IOT_KPRTL_LOCK_W

Field HOST_IOT_KPRTL_LOCK writer - This register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.