nrf52840_pac::ccm

Module mode

Operation mode

Structs

MODE_SPEC

Operation mode

R

Register MODE reader

W

Register MODE writer

Enums

DATARATE_A

Radio data rate that the CCM shall run synchronous with

LENGTH_A

Packet length configuration

MODE_A

The mode of operation to be used. The settings in this register apply whenever either the KSGEN or CRYPT tasks are triggered.

Type Aliases

DATARATE_R

Field DATARATE reader - Radio data rate that the CCM shall run synchronous with

DATARATE_W

Field DATARATE writer - Radio data rate that the CCM shall run synchronous with

LENGTH_R

Field LENGTH reader - Packet length configuration

LENGTH_W

Field LENGTH writer - Packet length configuration

MODE_R

Field MODE reader - The mode of operation to be used. The settings in this register apply whenever either the KSGEN or CRYPT tasks are triggered.

MODE_W

Field MODE writer - The mode of operation to be used. The settings in this register apply whenever either the KSGEN or CRYPT tasks are triggered.