nrf52840_pac::ccm::mode

Type Alias LENGTH_R

pub type LENGTH_R = BitReader<LENGTH_A>;

Field LENGTH reader - Packet length configuration

Aliased Type

struct LENGTH_R { /* private fields */ }

Implementations

impl LENGTH_R

pub fn variant(&self) -> LENGTH_A

Get enumerated values variant

pub fn is_default(&self) -> bool

Checks if the value of the field is DEFAULT

pub fn is_extended(&self) -> bool

Checks if the value of the field is EXTENDED