nrf52840_pac::ccm

Module scratchptr

Pointer to data area used for temporary storage

Structs

R

Register SCRATCHPTR reader

SCRATCHPTR_SPEC

Pointer to data area used for temporary storage

W

Register SCRATCHPTR writer

Type Aliases

SCRATCHPTR_R

Field SCRATCHPTR reader - Pointer to a scratch data area used for temporary storage during key-stream generation, MIC generation and encryption/decryption.

SCRATCHPTR_W

Field SCRATCHPTR writer - Pointer to a scratch data area used for temporary storage during key-stream generation, MIC generation and encryption/decryption.