nrf52840_pac::clock::lfclkstat

Type Alias STATE_R

pub type STATE_R = BitReader<STATE_A>;

Field STATE reader - LFCLK state

Aliased Type

struct STATE_R { /* private fields */ }

Implementations

impl STATE_R

pub fn variant(&self) -> STATE_A

Get enumerated values variant

pub fn is_not_running(&self) -> bool

Checks if the value of the field is NOT_RUNNING

pub fn is_running(&self) -> bool

Checks if the value of the field is RUNNING