nrf52840_pac::clock

Struct RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub tasks_hfclkstart: TASKS_HFCLKSTART,
    pub tasks_hfclkstop: TASKS_HFCLKSTOP,
    pub tasks_lfclkstart: TASKS_LFCLKSTART,
    pub tasks_lfclkstop: TASKS_LFCLKSTOP,
    pub tasks_cal: TASKS_CAL,
    pub tasks_ctstart: TASKS_CTSTART,
    pub tasks_ctstop: TASKS_CTSTOP,
    pub events_hfclkstarted: EVENTS_HFCLKSTARTED,
    pub events_lfclkstarted: EVENTS_LFCLKSTARTED,
    pub events_done: EVENTS_DONE,
    pub events_ctto: EVENTS_CTTO,
    pub events_ctstarted: EVENTS_CTSTARTED,
    pub events_ctstopped: EVENTS_CTSTOPPED,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub hfclkrun: HFCLKRUN,
    pub hfclkstat: HFCLKSTAT,
    pub lfclkrun: LFCLKRUN,
    pub lfclkstat: LFCLKSTAT,
    pub lfclksrccopy: LFCLKSRCCOPY,
    pub lfclksrc: LFCLKSRC,
    pub hfxodebounce: HFXODEBOUNCE,
    pub ctiv: CTIV,
    pub traceconfig: TRACECONFIG,
    pub lfrcmode: LFRCMODE,
    /* private fields */

}

Register block

Fields
tasks_hfclkstart: TASKS_HFCLKSTART

0x00 - Start HFXO crystal oscillator

tasks_hfclkstop: TASKS_HFCLKSTOP

0x04 - Stop HFXO crystal oscillator

tasks_lfclkstart: TASKS_LFCLKSTART

0x08 - Start LFCLK

tasks_lfclkstop: TASKS_LFCLKSTOP

0x0c - Stop LFCLK

tasks_cal: TASKS_CAL

0x10 - Start calibration of LFRC

tasks_ctstart: TASKS_CTSTART

0x14 - Start calibration timer

tasks_ctstop: TASKS_CTSTOP

0x18 - Stop calibration timer

events_hfclkstarted: EVENTS_HFCLKSTARTED

0x100 - HFXO crystal oscillator started

events_lfclkstarted: EVENTS_LFCLKSTARTED

0x104 - LFCLK started

events_done: EVENTS_DONE

0x10c - Calibration of LFRC completed

events_ctto: EVENTS_CTTO

0x110 - Calibration timer timeout

events_ctstarted: EVENTS_CTSTARTED

0x128 - Calibration timer has been started and is ready to process new tasks

events_ctstopped: EVENTS_CTSTOPPED

0x12c - Calibration timer has been stopped and is ready to process new tasks

intenset: INTENSET

0x304 - Enable interrupt

intenclr: INTENCLR

0x308 - Disable interrupt

hfclkrun: HFCLKRUN

0x408 - Status indicating that HFCLKSTART task has been triggered

hfclkstat: HFCLKSTAT

0x40c - HFCLK status

lfclkrun: LFCLKRUN

0x414 - Status indicating that LFCLKSTART task has been triggered

lfclkstat: LFCLKSTAT

0x418 - LFCLK status

lfclksrccopy: LFCLKSRCCOPY

0x41c - Copy of LFCLKSRC register, set when LFCLKSTART task was triggered

lfclksrc: LFCLKSRC

0x518 - Clock source for the LFCLK

hfxodebounce: HFXODEBOUNCE

0x528 - HFXO debounce time. The HFXO is started by triggering the TASKS_HFCLKSTART task.

ctiv: CTIV

0x538 - Calibration timer interval

traceconfig: TRACECONFIG

0x55c - Clocking options for the trace port debug interface

lfrcmode: LFRCMODE

0x5b4 - LFRC mode configuration