nrf52840_pac::clock::traceconfig

Type Alias TRACEMUX_R

pub type TRACEMUX_R = FieldReader<u8, TRACEMUX_A>;

Field TRACEMUX reader - Pin multiplexing of trace signals. See pin assignment chapter for more details.

Aliased Type

struct TRACEMUX_R { /* private fields */ }

Implementations

impl TRACEMUX_R

pub fn variant(&self) -> Option<TRACEMUX_A>

Get enumerated values variant

pub fn is_gpio(&self) -> bool

Checks if the value of the field is GPIO

pub fn is_serial(&self) -> bool

Checks if the value of the field is SERIAL

pub fn is_parallel(&self) -> bool

Checks if the value of the field is PARALLEL