nrf52840_pac::comp::extrefsel

Type Alias EXTREFSEL_R

pub type EXTREFSEL_R = FieldReader<u8, EXTREFSEL_A>;

Field EXTREFSEL reader - External analog reference select

Aliased Type

struct EXTREFSEL_R { /* private fields */ }

Implementations

impl EXTREFSEL_R

pub fn variant(&self) -> EXTREFSEL_A

Get enumerated values variant

pub fn is_analog_reference0(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE0

pub fn is_analog_reference1(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE1

pub fn is_analog_reference2(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE2

pub fn is_analog_reference3(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE3

pub fn is_analog_reference4(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE4

pub fn is_analog_reference5(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE5

pub fn is_analog_reference6(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE6

pub fn is_analog_reference7(&self) -> bool

Checks if the value of the field is ANALOG_REFERENCE7