nrf52840_pac::ficr::info::ram

Type Alias RAM_R

pub type RAM_R = FieldReader<u32, RAM_A>;

Field RAM reader - RAM variant

Aliased Type

struct RAM_R { /* private fields */ }

Implementations

impl RAM_R

pub fn variant(&self) -> Option<RAM_A>

Get enumerated values variant

pub fn is_k16(&self) -> bool

Checks if the value of the field is K16

pub fn is_k32(&self) -> bool

Checks if the value of the field is K32

pub fn is_k64(&self) -> bool

Checks if the value of the field is K64

pub fn is_k128(&self) -> bool

Checks if the value of the field is K128

pub fn is_k256(&self) -> bool

Checks if the value of the field is K256

pub fn is_unspecified(&self) -> bool

Checks if the value of the field is UNSPECIFIED