nrf52840_pac::ficr::info::variant

Type Alias VARIANT_R

pub type VARIANT_R = FieldReader<u32, VARIANT_A>;

Field VARIANT reader - Build code (hardware version and production configuration). Encoded as ASCII.

Aliased Type

struct VARIANT_R { /* private fields */ }

Implementations

impl VARIANT_R

pub fn variant(&self) -> Option<VARIANT_A>

Get enumerated values variant

pub fn is_aaaa(&self) -> bool

Checks if the value of the field is AAAA

pub fn is_baaa(&self) -> bool

Checks if the value of the field is BAAA

pub fn is_caaa(&self) -> bool

Checks if the value of the field is CAAA

pub fn is_aaba(&self) -> bool

Checks if the value of the field is AABA

pub fn is_aabb(&self) -> bool

Checks if the value of the field is AABB

pub fn is_aaca(&self) -> bool

Checks if the value of the field is AACA

pub fn is_aaab(&self) -> bool

Checks if the value of the field is AAAB

pub fn is_unspecified(&self) -> bool

Checks if the value of the field is UNSPECIFIED