nrf52840_pac::generic

Trait Readable

pub trait Readable: RegisterSpec {
    type Reader: From<R<Self>> + Deref<Target = R<Self>>;
}

Trait implemented by readable registers to enable the read method.

Registers marked with Writable can be also modify’ed.

Required Associated Types

type Reader: From<R<Self>> + Deref<Target = R<Self>>

Result from a call to read and argument to modify.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.