nrf52840_pac::generic

Trait Resettable

pub trait Resettable: RegisterSpec {
    // Required method
    fn reset_value() -> Self::Ux;
}

Reset value of the register.

This value is the initial value for the write method. It can also be directly written to the register by using the reset method.

Required Methods

fn reset_value() -> Self::Ux

Reset value of the register.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.