nrf52840_pac::i2s::config::format

Type Alias FORMAT_R

pub type FORMAT_R = BitReader<FORMAT_A>;

Field FORMAT reader - Frame format.

Aliased Type

struct FORMAT_R { /* private fields */ }

Implementations

impl FORMAT_R

pub fn variant(&self) -> FORMAT_A

Get enumerated values variant

pub fn is_i2s(&self) -> bool

Checks if the value of the field is I2S

pub fn is_aligned(&self) -> bool

Checks if the value of the field is ALIGNED