nrf52840_pac::i2s::config::ratio

Type Alias RATIO_R

pub type RATIO_R = FieldReader<u8, RATIO_A>;

Field RATIO reader - MCK / LRCK ratio.

Aliased Type

struct RATIO_R { /* private fields */ }

Implementations

impl RATIO_R

pub fn variant(&self) -> Option<RATIO_A>

Get enumerated values variant

pub fn is_32x(&self) -> bool

Checks if the value of the field is _32X

pub fn is_48x(&self) -> bool

Checks if the value of the field is _48X

pub fn is_64x(&self) -> bool

Checks if the value of the field is _64X

pub fn is_96x(&self) -> bool

Checks if the value of the field is _96X

pub fn is_128x(&self) -> bool

Checks if the value of the field is _128X

pub fn is_192x(&self) -> bool

Checks if the value of the field is _192X

pub fn is_256x(&self) -> bool

Checks if the value of the field is _256X

pub fn is_384x(&self) -> bool

Checks if the value of the field is _384X

pub fn is_512x(&self) -> bool

Checks if the value of the field is _512X