nrf52840_pac::i2s::config::swidth

Type Alias SWIDTH_R

pub type SWIDTH_R = FieldReader<u8, SWIDTH_A>;

Field SWIDTH reader - Sample width.

Aliased Type

struct SWIDTH_R { /* private fields */ }

Implementations

impl SWIDTH_R

pub fn variant(&self) -> Option<SWIDTH_A>

Get enumerated values variant

pub fn is_8bit(&self) -> bool

Checks if the value of the field is _8BIT

pub fn is_16bit(&self) -> bool

Checks if the value of the field is _16BIT

pub fn is_24bit(&self) -> bool

Checks if the value of the field is _24BIT