nrf52840_pac::i2s::rxd
Module ptr

Receive buffer RAM start address.

Structs

PTR_SPEC

Receive buffer RAM start address.

R

Register PTR reader

W

Register PTR writer

Type Aliases

PTR_R

Field PTR reader - Receive buffer Data RAM start address. When receiving, words containing samples will be written to this address. This address is a word aligned Data RAM address.

PTR_W

Field PTR writer - Receive buffer Data RAM start address. When receiving, words containing samples will be written to this address. This address is a word aligned Data RAM address.