Type Alias nrf52840_pac::lpcomp::psel::PSEL_R

pub type PSEL_R = FieldReader<u8, PSEL_A>;

Field PSEL reader - Analog pin select

Aliased Type

struct PSEL_R { /* private fields */ }

Implementations

impl PSEL_R

pub fn variant(&self) -> PSEL_A

Get enumerated values variant

pub fn is_analog_input0(&self) -> bool

Checks if the value of the field is ANALOG_INPUT0

pub fn is_analog_input1(&self) -> bool

Checks if the value of the field is ANALOG_INPUT1

pub fn is_analog_input2(&self) -> bool

Checks if the value of the field is ANALOG_INPUT2

pub fn is_analog_input3(&self) -> bool

Checks if the value of the field is ANALOG_INPUT3

pub fn is_analog_input4(&self) -> bool

Checks if the value of the field is ANALOG_INPUT4

pub fn is_analog_input5(&self) -> bool

Checks if the value of the field is ANALOG_INPUT5

pub fn is_analog_input6(&self) -> bool

Checks if the value of the field is ANALOG_INPUT6

pub fn is_analog_input7(&self) -> bool

Checks if the value of the field is ANALOG_INPUT7