Type Alias nrf52840_pac::lpcomp::refsel::REFSEL_R

pub type REFSEL_R = FieldReader<u8, REFSEL_A>;

Field REFSEL reader - Reference select

Aliased Type

struct REFSEL_R { /* private fields */ }

Implementations

impl REFSEL_R

pub fn variant(&self) -> REFSEL_A

Get enumerated values variant

pub fn is_ref1_8vdd(&self) -> bool

Checks if the value of the field is REF1_8VDD

pub fn is_ref2_8vdd(&self) -> bool

Checks if the value of the field is REF2_8VDD

pub fn is_ref3_8vdd(&self) -> bool

Checks if the value of the field is REF3_8VDD

pub fn is_ref4_8vdd(&self) -> bool

Checks if the value of the field is REF4_8VDD

pub fn is_ref5_8vdd(&self) -> bool

Checks if the value of the field is REF5_8VDD

pub fn is_ref6_8vdd(&self) -> bool

Checks if the value of the field is REF6_8VDD

pub fn is_ref7_8vdd(&self) -> bool

Checks if the value of the field is REF7_8VDD

pub fn is_aref(&self) -> bool

Checks if the value of the field is AREF

pub fn is_ref1_16vdd(&self) -> bool

Checks if the value of the field is REF1_16VDD

pub fn is_ref3_16vdd(&self) -> bool

Checks if the value of the field is REF3_16VDD

pub fn is_ref5_16vdd(&self) -> bool

Checks if the value of the field is REF5_16VDD

pub fn is_ref7_16vdd(&self) -> bool

Checks if the value of the field is REF7_16VDD

pub fn is_ref9_16vdd(&self) -> bool

Checks if the value of the field is REF9_16VDD

pub fn is_ref11_16vdd(&self) -> bool

Checks if the value of the field is REF11_16VDD

pub fn is_ref13_16vdd(&self) -> bool

Checks if the value of the field is REF13_16VDD

pub fn is_ref15_16vdd(&self) -> bool

Checks if the value of the field is REF15_16VDD