nrf52840_pac::mwu::nmien

Type Alias REGION2WA_R

pub type REGION2WA_R = BitReader<REGION2WA_A>;

Field REGION2WA reader - Enable or disable non-maskable interrupt for REGION[2].WA event

Aliased Type

pub struct REGION2WA_R { /* private fields */ }

Implementations

impl REGION2WA_R

pub fn variant(&self) -> REGION2WA_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_enabled(&self) -> bool

Checks if the value of the field is ENABLED