nrf52840_pac::mwu::perregion

Module substatra

Description cluster[n]: Source of event/interrupt in region n, read access detected while corresponding subregion was enabled for watching

Structs

R

Register SUBSTATRA reader

SUBSTATRA_SPEC

Description cluster[n]: Source of event/interrupt in region n, read access detected while corresponding subregion was enabled for watching

W

Register SUBSTATRA writer

Enums

SR0_A

Subregion 0 in region n (write ‘1’ to clear)

SR1_A

Subregion 1 in region n (write ‘1’ to clear)

SR2_A

Subregion 2 in region n (write ‘1’ to clear)

SR3_A

Subregion 3 in region n (write ‘1’ to clear)

SR4_A

Subregion 4 in region n (write ‘1’ to clear)

SR5_A

Subregion 5 in region n (write ‘1’ to clear)

SR6_A

Subregion 6 in region n (write ‘1’ to clear)

SR7_A

Subregion 7 in region n (write ‘1’ to clear)

SR8_A

Subregion 8 in region n (write ‘1’ to clear)

SR9_A

Subregion 9 in region n (write ‘1’ to clear)

SR10_A

Subregion 10 in region n (write ‘1’ to clear)

SR11_A

Subregion 11 in region n (write ‘1’ to clear)

SR12_A

Subregion 12 in region n (write ‘1’ to clear)

SR13_A

Subregion 13 in region n (write ‘1’ to clear)

SR14_A

Subregion 14 in region n (write ‘1’ to clear)

SR15_A

Subregion 15 in region n (write ‘1’ to clear)

SR16_A

Subregion 16 in region n (write ‘1’ to clear)

SR17_A

Subregion 17 in region n (write ‘1’ to clear)

SR18_A

Subregion 18 in region n (write ‘1’ to clear)

SR19_A

Subregion 19 in region n (write ‘1’ to clear)

SR20_A

Subregion 20 in region n (write ‘1’ to clear)

SR21_A

Subregion 21 in region n (write ‘1’ to clear)

SR22_A

Subregion 22 in region n (write ‘1’ to clear)

SR23_A

Subregion 23 in region n (write ‘1’ to clear)

SR24_A

Subregion 24 in region n (write ‘1’ to clear)

SR25_A

Subregion 25 in region n (write ‘1’ to clear)

SR26_A

Subregion 26 in region n (write ‘1’ to clear)

SR27_A

Subregion 27 in region n (write ‘1’ to clear)

SR28_A

Subregion 28 in region n (write ‘1’ to clear)

SR29_A

Subregion 29 in region n (write ‘1’ to clear)

SR30_A

Subregion 30 in region n (write ‘1’ to clear)

SR31_A

Subregion 31 in region n (write ‘1’ to clear)

Type Aliases

SR0_R

Field SR0 reader - Subregion 0 in region n (write ‘1’ to clear)

SR0_W

Field SR0 writer - Subregion 0 in region n (write ‘1’ to clear)

SR1_R

Field SR1 reader - Subregion 1 in region n (write ‘1’ to clear)

SR1_W

Field SR1 writer - Subregion 1 in region n (write ‘1’ to clear)

SR2_R

Field SR2 reader - Subregion 2 in region n (write ‘1’ to clear)

SR2_W

Field SR2 writer - Subregion 2 in region n (write ‘1’ to clear)

SR3_R

Field SR3 reader - Subregion 3 in region n (write ‘1’ to clear)

SR3_W

Field SR3 writer - Subregion 3 in region n (write ‘1’ to clear)

SR4_R

Field SR4 reader - Subregion 4 in region n (write ‘1’ to clear)

SR4_W

Field SR4 writer - Subregion 4 in region n (write ‘1’ to clear)

SR5_R

Field SR5 reader - Subregion 5 in region n (write ‘1’ to clear)

SR5_W

Field SR5 writer - Subregion 5 in region n (write ‘1’ to clear)

SR6_R

Field SR6 reader - Subregion 6 in region n (write ‘1’ to clear)

SR6_W

Field SR6 writer - Subregion 6 in region n (write ‘1’ to clear)

SR7_R

Field SR7 reader - Subregion 7 in region n (write ‘1’ to clear)

SR7_W

Field SR7 writer - Subregion 7 in region n (write ‘1’ to clear)

SR8_R

Field SR8 reader - Subregion 8 in region n (write ‘1’ to clear)

SR8_W

Field SR8 writer - Subregion 8 in region n (write ‘1’ to clear)

SR9_R

Field SR9 reader - Subregion 9 in region n (write ‘1’ to clear)

SR9_W

Field SR9 writer - Subregion 9 in region n (write ‘1’ to clear)

SR10_R

Field SR10 reader - Subregion 10 in region n (write ‘1’ to clear)

SR10_W

Field SR10 writer - Subregion 10 in region n (write ‘1’ to clear)

SR11_R

Field SR11 reader - Subregion 11 in region n (write ‘1’ to clear)

SR11_W

Field SR11 writer - Subregion 11 in region n (write ‘1’ to clear)

SR12_R

Field SR12 reader - Subregion 12 in region n (write ‘1’ to clear)

SR12_W

Field SR12 writer - Subregion 12 in region n (write ‘1’ to clear)

SR13_R

Field SR13 reader - Subregion 13 in region n (write ‘1’ to clear)

SR13_W

Field SR13 writer - Subregion 13 in region n (write ‘1’ to clear)

SR14_R

Field SR14 reader - Subregion 14 in region n (write ‘1’ to clear)

SR14_W

Field SR14 writer - Subregion 14 in region n (write ‘1’ to clear)

SR15_R

Field SR15 reader - Subregion 15 in region n (write ‘1’ to clear)

SR15_W

Field SR15 writer - Subregion 15 in region n (write ‘1’ to clear)

SR16_R

Field SR16 reader - Subregion 16 in region n (write ‘1’ to clear)

SR16_W

Field SR16 writer - Subregion 16 in region n (write ‘1’ to clear)

SR17_R

Field SR17 reader - Subregion 17 in region n (write ‘1’ to clear)

SR17_W

Field SR17 writer - Subregion 17 in region n (write ‘1’ to clear)

SR18_R

Field SR18 reader - Subregion 18 in region n (write ‘1’ to clear)

SR18_W

Field SR18 writer - Subregion 18 in region n (write ‘1’ to clear)

SR19_R

Field SR19 reader - Subregion 19 in region n (write ‘1’ to clear)

SR19_W

Field SR19 writer - Subregion 19 in region n (write ‘1’ to clear)

SR20_R

Field SR20 reader - Subregion 20 in region n (write ‘1’ to clear)

SR20_W

Field SR20 writer - Subregion 20 in region n (write ‘1’ to clear)

SR21_R

Field SR21 reader - Subregion 21 in region n (write ‘1’ to clear)

SR21_W

Field SR21 writer - Subregion 21 in region n (write ‘1’ to clear)

SR22_R

Field SR22 reader - Subregion 22 in region n (write ‘1’ to clear)

SR22_W

Field SR22 writer - Subregion 22 in region n (write ‘1’ to clear)

SR23_R

Field SR23 reader - Subregion 23 in region n (write ‘1’ to clear)

SR23_W

Field SR23 writer - Subregion 23 in region n (write ‘1’ to clear)

SR24_R

Field SR24 reader - Subregion 24 in region n (write ‘1’ to clear)

SR24_W

Field SR24 writer - Subregion 24 in region n (write ‘1’ to clear)

SR25_R

Field SR25 reader - Subregion 25 in region n (write ‘1’ to clear)

SR25_W

Field SR25 writer - Subregion 25 in region n (write ‘1’ to clear)

SR26_R

Field SR26 reader - Subregion 26 in region n (write ‘1’ to clear)

SR26_W

Field SR26 writer - Subregion 26 in region n (write ‘1’ to clear)

SR27_R

Field SR27 reader - Subregion 27 in region n (write ‘1’ to clear)

SR27_W

Field SR27 writer - Subregion 27 in region n (write ‘1’ to clear)

SR28_R

Field SR28 reader - Subregion 28 in region n (write ‘1’ to clear)

SR28_W

Field SR28 writer - Subregion 28 in region n (write ‘1’ to clear)

SR29_R

Field SR29 reader - Subregion 29 in region n (write ‘1’ to clear)

SR29_W

Field SR29 writer - Subregion 29 in region n (write ‘1’ to clear)

SR30_R

Field SR30 reader - Subregion 30 in region n (write ‘1’ to clear)

SR30_W

Field SR30 writer - Subregion 30 in region n (write ‘1’ to clear)

SR31_R

Field SR31 reader - Subregion 31 in region n (write ‘1’ to clear)

SR31_W

Field SR31 writer - Subregion 31 in region n (write ‘1’ to clear)