nrf52840_pac::mwu::perregion::substatwa
Type Alias SR11_R

pub type SR11_R = BitReader<SR11_A>;

Field SR11 reader - Subregion 11 in region n (write ‘1’ to clear)

Aliased Type

struct SR11_R { /* private fields */ }

Implementations

impl SR11_R

pub fn variant(&self) -> SR11_A

Get enumerated values variant

pub fn is_no_access(&self) -> bool

Checks if the value of the field is NO_ACCESS

pub fn is_access(&self) -> bool

Checks if the value of the field is ACCESS