Module nrf52840_pac::nfct

NFC-A compatible radio

Re-exports

pub use framestatus::FRAMESTATUS;

pub use txd::TXD;

pub use rxd::RXD;

Modules

autocolresconfig

Controls the auto collision resolution function. This setting must be done before the NFCT peripheral is enabled.

errorstatus

NFC Error Status register

events_autocolresstarted

Auto collision resolution process has started

events_collision

NFC auto collision resolution error reported.

events_endrx

RX buffer (as defined by PACKETPTR and MAXLEN) in Data RAM full.

events_endtx

Transmission of data in RAM has ended, and EasyDMA has ended accessing the TX buffer

events_error

NFC error reported. The ERRORSTATUS register contains details on the source of the error.

events_fielddetected

Remote NFC field detected

events_fieldlost

Remote NFC field lost

events_ready

The NFCT peripheral is ready to receive and send frames

events_rxerror

NFC RX frame error reported. The FRAMESTATUS.RX register contains details on the source of the error.

events_rxframeend

Received data has been checked (CRC, parity) and transferred to RAM, and EasyDMA has ended accessing the RX buffer

events_rxframestart

Marks the end of the first symbol of a received frame

events_selected

NFC auto collision resolution successfully completed

events_started

EasyDMA is ready to receive or send frames.

events_txframeend

Marks the end of the last transmitted on-air symbol of a frame

events_txframestart

Marks the start of the first symbol of a transmitted frame

fieldpresent

Indicates the presence or not of a valid field

framedelaymax

Maximum frame delay

framedelaymin

Minimum frame delay

framedelaymode

Configuration register for the Frame Delay Timer

framestatus

Cluster Unspecified

inten

Enable or disable interrupt

intenclr

Disable interrupt

intenset

Enable interrupt

maxlen

Size of the RAM buffer allocated to TXD and RXD data storage each

nfcid1_2nd_last

Second last NFCID1 part (7 or 10 bytes ID)

nfcid1_3rd_last

Third last NFCID1 part (10 bytes ID)

nfcid1_last

Last NFCID1 part (4, 7 or 10 bytes ID)

nfctagstate

NfcTag state register

packetptr

Packet pointer for TXD and RXD data storage in Data RAM

rxd

Cluster Unspecified

selres

NFC-A SEL_RES auto-response settings

sensres

NFC-A SENS_RES auto-response settings

shorts

Shortcut register

sleepstate

Sleep state during automatic collision resolution

tasks_activate

Activate NFCT peripheral for incoming and outgoing frames, change state to activated

tasks_disable

Disable NFCT peripheral

tasks_enablerxdata

Initializes the EasyDMA for receive.

tasks_goidle

Force state machine to IDLE state

tasks_gosleep

Force state machine to SLEEP_A state

tasks_sense

Enable NFC sense field mode, change state to sense mode

tasks_starttx

Start transmission of an outgoing frame, change state to transmit

txd

Cluster Unspecified

Structs

RegisterBlock

Register block

Type Aliases

AUTOCOLRESCONFIG

AUTOCOLRESCONFIG (rw) register accessor: an alias for Reg<AUTOCOLRESCONFIG_SPEC>

ERRORSTATUS

ERRORSTATUS (rw) register accessor: an alias for Reg<ERRORSTATUS_SPEC>

EVENTS_AUTOCOLRESSTARTED

EVENTS_AUTOCOLRESSTARTED (rw) register accessor: an alias for Reg<EVENTS_AUTOCOLRESSTARTED_SPEC>

EVENTS_COLLISION

EVENTS_COLLISION (rw) register accessor: an alias for Reg<EVENTS_COLLISION_SPEC>

EVENTS_ENDRX

EVENTS_ENDRX (rw) register accessor: an alias for Reg<EVENTS_ENDRX_SPEC>

EVENTS_ENDTX

EVENTS_ENDTX (rw) register accessor: an alias for Reg<EVENTS_ENDTX_SPEC>

EVENTS_ERROR

EVENTS_ERROR (rw) register accessor: an alias for Reg<EVENTS_ERROR_SPEC>

EVENTS_FIELDDETECTED

EVENTS_FIELDDETECTED (rw) register accessor: an alias for Reg<EVENTS_FIELDDETECTED_SPEC>

EVENTS_FIELDLOST

EVENTS_FIELDLOST (rw) register accessor: an alias for Reg<EVENTS_FIELDLOST_SPEC>

EVENTS_READY

EVENTS_READY (rw) register accessor: an alias for Reg<EVENTS_READY_SPEC>

EVENTS_RXERROR

EVENTS_RXERROR (rw) register accessor: an alias for Reg<EVENTS_RXERROR_SPEC>

EVENTS_RXFRAMEEND

EVENTS_RXFRAMEEND (rw) register accessor: an alias for Reg<EVENTS_RXFRAMEEND_SPEC>

EVENTS_RXFRAMESTART

EVENTS_RXFRAMESTART (rw) register accessor: an alias for Reg<EVENTS_RXFRAMESTART_SPEC>

EVENTS_SELECTED

EVENTS_SELECTED (rw) register accessor: an alias for Reg<EVENTS_SELECTED_SPEC>

EVENTS_STARTED

EVENTS_STARTED (rw) register accessor: an alias for Reg<EVENTS_STARTED_SPEC>

EVENTS_TXFRAMEEND

EVENTS_TXFRAMEEND (rw) register accessor: an alias for Reg<EVENTS_TXFRAMEEND_SPEC>

EVENTS_TXFRAMESTART

EVENTS_TXFRAMESTART (rw) register accessor: an alias for Reg<EVENTS_TXFRAMESTART_SPEC>

FIELDPRESENT

FIELDPRESENT (r) register accessor: an alias for Reg<FIELDPRESENT_SPEC>

FRAMEDELAYMAX

FRAMEDELAYMAX (rw) register accessor: an alias for Reg<FRAMEDELAYMAX_SPEC>

FRAMEDELAYMIN

FRAMEDELAYMIN (rw) register accessor: an alias for Reg<FRAMEDELAYMIN_SPEC>

FRAMEDELAYMODE

FRAMEDELAYMODE (rw) register accessor: an alias for Reg<FRAMEDELAYMODE_SPEC>

INTEN

INTEN (rw) register accessor: an alias for Reg<INTEN_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

MAXLEN

MAXLEN (rw) register accessor: an alias for Reg<MAXLEN_SPEC>

NFCID1_2ND_LAST

NFCID1_2ND_LAST (rw) register accessor: an alias for Reg<NFCID1_2ND_LAST_SPEC>

NFCID1_3RD_LAST

NFCID1_3RD_LAST (rw) register accessor: an alias for Reg<NFCID1_3RD_LAST_SPEC>

NFCID1_LAST

NFCID1_LAST (rw) register accessor: an alias for Reg<NFCID1_LAST_SPEC>

NFCTAGSTATE

NFCTAGSTATE (r) register accessor: an alias for Reg<NFCTAGSTATE_SPEC>

PACKETPTR

PACKETPTR (rw) register accessor: an alias for Reg<PACKETPTR_SPEC>

SELRES

SELRES (rw) register accessor: an alias for Reg<SELRES_SPEC>

SENSRES

SENSRES (rw) register accessor: an alias for Reg<SENSRES_SPEC>

SHORTS

SHORTS (rw) register accessor: an alias for Reg<SHORTS_SPEC>

SLEEPSTATE

SLEEPSTATE (r) register accessor: an alias for Reg<SLEEPSTATE_SPEC>

TASKS_ACTIVATE

TASKS_ACTIVATE (w) register accessor: an alias for Reg<TASKS_ACTIVATE_SPEC>

TASKS_DISABLE

TASKS_DISABLE (w) register accessor: an alias for Reg<TASKS_DISABLE_SPEC>

TASKS_ENABLERXDATA

TASKS_ENABLERXDATA (w) register accessor: an alias for Reg<TASKS_ENABLERXDATA_SPEC>

TASKS_GOIDLE

TASKS_GOIDLE (w) register accessor: an alias for Reg<TASKS_GOIDLE_SPEC>

TASKS_GOSLEEP

TASKS_GOSLEEP (w) register accessor: an alias for Reg<TASKS_GOSLEEP_SPEC>

TASKS_SENSE

TASKS_SENSE (w) register accessor: an alias for Reg<TASKS_SENSE_SPEC>

TASKS_STARTTX

TASKS_STARTTX (w) register accessor: an alias for Reg<TASKS_STARTTX_SPEC>