Module nrf52840_pac::nfct::fieldpresent

Indicates the presence or not of a valid field

Structs

FIELDPRESENT_SPEC

Indicates the presence or not of a valid field

R

Register FIELDPRESENT reader

Enums

FIELDPRESENT_A

Indicates if a valid field is present. Available only in the activated state.

LOCKDETECT_A

Indicates if the low level has locked to the field

Type Aliases

FIELDPRESENT_R

Field FIELDPRESENT reader - Indicates if a valid field is present. Available only in the activated state.

LOCKDETECT_R

Field LOCKDETECT reader - Indicates if the low level has locked to the field