Type Alias nrf52840_pac::nfct::framedelaymode::FRAMEDELAYMODE_R

pub type FRAMEDELAYMODE_R = FieldReader<u8, FRAMEDELAYMODE_A>;

Field FRAMEDELAYMODE reader - Configuration register for the Frame Delay Timer

Aliased Type

struct FRAMEDELAYMODE_R { /* private fields */ }

Implementations

impl FRAMEDELAYMODE_R

pub fn variant(&self) -> FRAMEDELAYMODE_A

Get enumerated values variant

pub fn is_free_run(&self) -> bool

Checks if the value of the field is FREE_RUN

pub fn is_window(&self) -> bool

Checks if the value of the field is WINDOW

pub fn is_exact_val(&self) -> bool

Checks if the value of the field is EXACT_VAL

pub fn is_window_grid(&self) -> bool

Checks if the value of the field is WINDOW_GRID