nrf52840_pac::nfct::framestatus::rx

Type Alias CRCERROR_R

pub type CRCERROR_R = BitReader<CRCERROR_A>;

Field CRCERROR reader - No valid end of frame (EoF) detected

Aliased Type

pub struct CRCERROR_R { /* private fields */ }

Implementations

impl CRCERROR_R

pub fn variant(&self) -> CRCERROR_A

Get enumerated values variant

pub fn is_crccorrect(&self) -> bool

Checks if the value of the field is CRCCORRECT

pub fn is_crcerror(&self) -> bool

Checks if the value of the field is CRCERROR