Module nrf52840_pac::nfct::sensres

NFC-A SENS_RES auto-response settings

Structs

R

Register SENSRES reader

SENSRES_SPEC

NFC-A SENS_RES auto-response settings

W

Register SENSRES writer

Enums

BITFRAMESDD_A

Bit frame SDD as defined by the b5:b1 of byte 1 in SENS_RES response in the NFC Forum, NFC Digital Protocol Technical Specification

NFCIDSIZE_A

NFCID1 size. This value is used by the auto collision resolution engine.

Type Aliases

BITFRAMESDD_R

Field BITFRAMESDD reader - Bit frame SDD as defined by the b5:b1 of byte 1 in SENS_RES response in the NFC Forum, NFC Digital Protocol Technical Specification

BITFRAMESDD_W

Field BITFRAMESDD writer - Bit frame SDD as defined by the b5:b1 of byte 1 in SENS_RES response in the NFC Forum, NFC Digital Protocol Technical Specification

NFCIDSIZE_R

Field NFCIDSIZE reader - NFCID1 size. This value is used by the auto collision resolution engine.

NFCIDSIZE_W

Field NFCIDSIZE writer - NFCID1 size. This value is used by the auto collision resolution engine.

PLATFCONFIG_R

Field PLATFCONFIG reader - Tag platform configuration as defined by the b4:b1 of byte 2 in SENS_RES response in the NFC Forum, NFC Digital Protocol Technical Specification

PLATFCONFIG_W

Field PLATFCONFIG writer - Tag platform configuration as defined by the b4:b1 of byte 2 in SENS_RES response in the NFC Forum, NFC Digital Protocol Technical Specification

RFU5_R

Field RFU5 reader - Reserved for future use. Shall be 0.

RFU5_W

Field RFU5 writer - Reserved for future use. Shall be 0.

RFU74_R

Field RFU74 reader - Reserved for future use. Shall be 0.

RFU74_W

Field RFU74 writer - Reserved for future use. Shall be 0.