nrf52840_pac::nfct::sensres

Type Alias NFCIDSIZE_R

pub type NFCIDSIZE_R = FieldReader<u8, NFCIDSIZE_A>;

Field NFCIDSIZE reader - NFCID1 size. This value is used by the auto collision resolution engine.

Aliased Type

struct NFCIDSIZE_R { /* private fields */ }

Implementations

impl NFCIDSIZE_R

pub fn variant(&self) -> Option<NFCIDSIZE_A>

Get enumerated values variant

pub fn is_nfcid1single(&self) -> bool

Checks if the value of the field is NFCID1SINGLE

pub fn is_nfcid1double(&self) -> bool

Checks if the value of the field is NFCID1DOUBLE

pub fn is_nfcid1triple(&self) -> bool

Checks if the value of the field is NFCID1TRIPLE