nrf52840_pac::nfct

Struct RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub tasks_activate: TASKS_ACTIVATE,
    pub tasks_disable: TASKS_DISABLE,
    pub tasks_sense: TASKS_SENSE,
    pub tasks_starttx: TASKS_STARTTX,
    pub tasks_enablerxdata: TASKS_ENABLERXDATA,
    pub tasks_goidle: TASKS_GOIDLE,
    pub tasks_gosleep: TASKS_GOSLEEP,
    pub events_ready: EVENTS_READY,
    pub events_fielddetected: EVENTS_FIELDDETECTED,
    pub events_fieldlost: EVENTS_FIELDLOST,
    pub events_txframestart: EVENTS_TXFRAMESTART,
    pub events_txframeend: EVENTS_TXFRAMEEND,
    pub events_rxframestart: EVENTS_RXFRAMESTART,
    pub events_rxframeend: EVENTS_RXFRAMEEND,
    pub events_error: EVENTS_ERROR,
    pub events_rxerror: EVENTS_RXERROR,
    pub events_endrx: EVENTS_ENDRX,
    pub events_endtx: EVENTS_ENDTX,
    pub events_autocolresstarted: EVENTS_AUTOCOLRESSTARTED,
    pub events_collision: EVENTS_COLLISION,
    pub events_selected: EVENTS_SELECTED,
    pub events_started: EVENTS_STARTED,
    pub shorts: SHORTS,
    pub inten: INTEN,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub errorstatus: ERRORSTATUS,
    pub framestatus: FRAMESTATUS,
    pub nfctagstate: NFCTAGSTATE,
    pub sleepstate: SLEEPSTATE,
    pub fieldpresent: FIELDPRESENT,
    pub framedelaymin: FRAMEDELAYMIN,
    pub framedelaymax: FRAMEDELAYMAX,
    pub framedelaymode: FRAMEDELAYMODE,
    pub packetptr: PACKETPTR,
    pub maxlen: MAXLEN,
    pub txd: TXD,
    pub rxd: RXD,
    pub nfcid1_last: NFCID1_LAST,
    pub nfcid1_2nd_last: NFCID1_2ND_LAST,
    pub nfcid1_3rd_last: NFCID1_3RD_LAST,
    pub autocolresconfig: AUTOCOLRESCONFIG,
    pub sensres: SENSRES,
    pub selres: SELRES,
    /* private fields */

}

Register block

Fields
tasks_activate: TASKS_ACTIVATE

0x00 - Activate NFCT peripheral for incoming and outgoing frames, change state to activated

tasks_disable: TASKS_DISABLE

0x04 - Disable NFCT peripheral

tasks_sense: TASKS_SENSE

0x08 - Enable NFC sense field mode, change state to sense mode

tasks_starttx: TASKS_STARTTX

0x0c - Start transmission of an outgoing frame, change state to transmit

tasks_enablerxdata: TASKS_ENABLERXDATA

0x1c - Initializes the EasyDMA for receive.

tasks_goidle: TASKS_GOIDLE

0x24 - Force state machine to IDLE state

tasks_gosleep: TASKS_GOSLEEP

0x28 - Force state machine to SLEEP_A state

events_ready: EVENTS_READY

0x100 - The NFCT peripheral is ready to receive and send frames

events_fielddetected: EVENTS_FIELDDETECTED

0x104 - Remote NFC field detected

events_fieldlost: EVENTS_FIELDLOST

0x108 - Remote NFC field lost

events_txframestart: EVENTS_TXFRAMESTART

0x10c - Marks the start of the first symbol of a transmitted frame

events_txframeend: EVENTS_TXFRAMEEND

0x110 - Marks the end of the last transmitted on-air symbol of a frame

events_rxframestart: EVENTS_RXFRAMESTART

0x114 - Marks the end of the first symbol of a received frame

events_rxframeend: EVENTS_RXFRAMEEND

0x118 - Received data has been checked (CRC, parity) and transferred to RAM, and EasyDMA has ended accessing the RX buffer

events_error: EVENTS_ERROR

0x11c - NFC error reported. The ERRORSTATUS register contains details on the source of the error.

events_rxerror: EVENTS_RXERROR

0x128 - NFC RX frame error reported. The FRAMESTATUS.RX register contains details on the source of the error.

events_endrx: EVENTS_ENDRX

0x12c - RX buffer (as defined by PACKETPTR and MAXLEN) in Data RAM full.

events_endtx: EVENTS_ENDTX

0x130 - Transmission of data in RAM has ended, and EasyDMA has ended accessing the TX buffer

events_autocolresstarted: EVENTS_AUTOCOLRESSTARTED

0x138 - Auto collision resolution process has started

events_collision: EVENTS_COLLISION

0x148 - NFC auto collision resolution error reported.

events_selected: EVENTS_SELECTED

0x14c - NFC auto collision resolution successfully completed

events_started: EVENTS_STARTED

0x150 - EasyDMA is ready to receive or send frames.

shorts: SHORTS

0x200 - Shortcut register

inten: INTEN

0x300 - Enable or disable interrupt

intenset: INTENSET

0x304 - Enable interrupt

intenclr: INTENCLR

0x308 - Disable interrupt

errorstatus: ERRORSTATUS

0x404 - NFC Error Status register

framestatus: FRAMESTATUS

0x40c - Unspecified

nfctagstate: NFCTAGSTATE

0x410 - NfcTag state register

sleepstate: SLEEPSTATE

0x420 - Sleep state during automatic collision resolution

fieldpresent: FIELDPRESENT

0x43c - Indicates the presence or not of a valid field

framedelaymin: FRAMEDELAYMIN

0x504 - Minimum frame delay

framedelaymax: FRAMEDELAYMAX

0x508 - Maximum frame delay

framedelaymode: FRAMEDELAYMODE

0x50c - Configuration register for the Frame Delay Timer

packetptr: PACKETPTR

0x510 - Packet pointer for TXD and RXD data storage in Data RAM

maxlen: MAXLEN

0x514 - Size of the RAM buffer allocated to TXD and RXD data storage each

txd: TXD

0x518..0x520 - Unspecified

rxd: RXD

0x520..0x528 - Unspecified

nfcid1_last: NFCID1_LAST

0x590 - Last NFCID1 part (4, 7 or 10 bytes ID)

nfcid1_2nd_last: NFCID1_2ND_LAST

0x594 - Second last NFCID1 part (7 or 10 bytes ID)

nfcid1_3rd_last: NFCID1_3RD_LAST

0x598 - Third last NFCID1 part (10 bytes ID)

autocolresconfig: AUTOCOLRESCONFIG

0x59c - Controls the auto collision resolution function. This setting must be done before the NFCT peripheral is enabled.

sensres: SENSRES

0x5a0 - NFC-A SENS_RES auto-response settings

selres: SELRES

0x5a4 - NFC-A SEL_RES auto-response settings