nrf52840_pac::p0

Module latch

Latch register indicating what GPIO pins that have met the criteria set in the PIN_CNF[n].SENSE registers

Structs

LATCH_SPEC

Latch register indicating what GPIO pins that have met the criteria set in the PIN_CNF[n].SENSE registers

R

Register LATCH reader

W

Register LATCH writer

Enums

PIN0_A

Status on whether PIN0 has met criteria set in PIN_CNF0.SENSE register. Write ‘1’ to clear.

PIN1_A

Status on whether PIN1 has met criteria set in PIN_CNF1.SENSE register. Write ‘1’ to clear.

PIN2_A

Status on whether PIN2 has met criteria set in PIN_CNF2.SENSE register. Write ‘1’ to clear.

PIN3_A

Status on whether PIN3 has met criteria set in PIN_CNF3.SENSE register. Write ‘1’ to clear.

PIN4_A

Status on whether PIN4 has met criteria set in PIN_CNF4.SENSE register. Write ‘1’ to clear.

PIN5_A

Status on whether PIN5 has met criteria set in PIN_CNF5.SENSE register. Write ‘1’ to clear.

PIN6_A

Status on whether PIN6 has met criteria set in PIN_CNF6.SENSE register. Write ‘1’ to clear.

PIN7_A

Status on whether PIN7 has met criteria set in PIN_CNF7.SENSE register. Write ‘1’ to clear.

PIN8_A

Status on whether PIN8 has met criteria set in PIN_CNF8.SENSE register. Write ‘1’ to clear.

PIN9_A

Status on whether PIN9 has met criteria set in PIN_CNF9.SENSE register. Write ‘1’ to clear.

PIN10_A

Status on whether PIN10 has met criteria set in PIN_CNF10.SENSE register. Write ‘1’ to clear.

PIN11_A

Status on whether PIN11 has met criteria set in PIN_CNF11.SENSE register. Write ‘1’ to clear.

PIN12_A

Status on whether PIN12 has met criteria set in PIN_CNF12.SENSE register. Write ‘1’ to clear.

PIN13_A

Status on whether PIN13 has met criteria set in PIN_CNF13.SENSE register. Write ‘1’ to clear.

PIN14_A

Status on whether PIN14 has met criteria set in PIN_CNF14.SENSE register. Write ‘1’ to clear.

PIN15_A

Status on whether PIN15 has met criteria set in PIN_CNF15.SENSE register. Write ‘1’ to clear.

PIN16_A

Status on whether PIN16 has met criteria set in PIN_CNF16.SENSE register. Write ‘1’ to clear.

PIN17_A

Status on whether PIN17 has met criteria set in PIN_CNF17.SENSE register. Write ‘1’ to clear.

PIN18_A

Status on whether PIN18 has met criteria set in PIN_CNF18.SENSE register. Write ‘1’ to clear.

PIN19_A

Status on whether PIN19 has met criteria set in PIN_CNF19.SENSE register. Write ‘1’ to clear.

PIN20_A

Status on whether PIN20 has met criteria set in PIN_CNF20.SENSE register. Write ‘1’ to clear.

PIN21_A

Status on whether PIN21 has met criteria set in PIN_CNF21.SENSE register. Write ‘1’ to clear.

PIN22_A

Status on whether PIN22 has met criteria set in PIN_CNF22.SENSE register. Write ‘1’ to clear.

PIN23_A

Status on whether PIN23 has met criteria set in PIN_CNF23.SENSE register. Write ‘1’ to clear.

PIN24_A

Status on whether PIN24 has met criteria set in PIN_CNF24.SENSE register. Write ‘1’ to clear.

PIN25_A

Status on whether PIN25 has met criteria set in PIN_CNF25.SENSE register. Write ‘1’ to clear.

PIN26_A

Status on whether PIN26 has met criteria set in PIN_CNF26.SENSE register. Write ‘1’ to clear.

PIN27_A

Status on whether PIN27 has met criteria set in PIN_CNF27.SENSE register. Write ‘1’ to clear.

PIN28_A

Status on whether PIN28 has met criteria set in PIN_CNF28.SENSE register. Write ‘1’ to clear.

PIN29_A

Status on whether PIN29 has met criteria set in PIN_CNF29.SENSE register. Write ‘1’ to clear.

PIN30_A

Status on whether PIN30 has met criteria set in PIN_CNF30.SENSE register. Write ‘1’ to clear.

PIN31_A

Status on whether PIN31 has met criteria set in PIN_CNF31.SENSE register. Write ‘1’ to clear.

Type Aliases

PIN0_R

Field PIN0 reader - Status on whether PIN0 has met criteria set in PIN_CNF0.SENSE register. Write ‘1’ to clear.

PIN0_W

Field PIN0 writer - Status on whether PIN0 has met criteria set in PIN_CNF0.SENSE register. Write ‘1’ to clear.

PIN1_R

Field PIN1 reader - Status on whether PIN1 has met criteria set in PIN_CNF1.SENSE register. Write ‘1’ to clear.

PIN1_W

Field PIN1 writer - Status on whether PIN1 has met criteria set in PIN_CNF1.SENSE register. Write ‘1’ to clear.

PIN2_R

Field PIN2 reader - Status on whether PIN2 has met criteria set in PIN_CNF2.SENSE register. Write ‘1’ to clear.

PIN2_W

Field PIN2 writer - Status on whether PIN2 has met criteria set in PIN_CNF2.SENSE register. Write ‘1’ to clear.

PIN3_R

Field PIN3 reader - Status on whether PIN3 has met criteria set in PIN_CNF3.SENSE register. Write ‘1’ to clear.

PIN3_W

Field PIN3 writer - Status on whether PIN3 has met criteria set in PIN_CNF3.SENSE register. Write ‘1’ to clear.

PIN4_R

Field PIN4 reader - Status on whether PIN4 has met criteria set in PIN_CNF4.SENSE register. Write ‘1’ to clear.

PIN4_W

Field PIN4 writer - Status on whether PIN4 has met criteria set in PIN_CNF4.SENSE register. Write ‘1’ to clear.

PIN5_R

Field PIN5 reader - Status on whether PIN5 has met criteria set in PIN_CNF5.SENSE register. Write ‘1’ to clear.

PIN5_W

Field PIN5 writer - Status on whether PIN5 has met criteria set in PIN_CNF5.SENSE register. Write ‘1’ to clear.

PIN6_R

Field PIN6 reader - Status on whether PIN6 has met criteria set in PIN_CNF6.SENSE register. Write ‘1’ to clear.

PIN6_W

Field PIN6 writer - Status on whether PIN6 has met criteria set in PIN_CNF6.SENSE register. Write ‘1’ to clear.

PIN7_R

Field PIN7 reader - Status on whether PIN7 has met criteria set in PIN_CNF7.SENSE register. Write ‘1’ to clear.

PIN7_W

Field PIN7 writer - Status on whether PIN7 has met criteria set in PIN_CNF7.SENSE register. Write ‘1’ to clear.

PIN8_R

Field PIN8 reader - Status on whether PIN8 has met criteria set in PIN_CNF8.SENSE register. Write ‘1’ to clear.

PIN8_W

Field PIN8 writer - Status on whether PIN8 has met criteria set in PIN_CNF8.SENSE register. Write ‘1’ to clear.

PIN9_R

Field PIN9 reader - Status on whether PIN9 has met criteria set in PIN_CNF9.SENSE register. Write ‘1’ to clear.

PIN9_W

Field PIN9 writer - Status on whether PIN9 has met criteria set in PIN_CNF9.SENSE register. Write ‘1’ to clear.

PIN10_R

Field PIN10 reader - Status on whether PIN10 has met criteria set in PIN_CNF10.SENSE register. Write ‘1’ to clear.

PIN10_W

Field PIN10 writer - Status on whether PIN10 has met criteria set in PIN_CNF10.SENSE register. Write ‘1’ to clear.

PIN11_R

Field PIN11 reader - Status on whether PIN11 has met criteria set in PIN_CNF11.SENSE register. Write ‘1’ to clear.

PIN11_W

Field PIN11 writer - Status on whether PIN11 has met criteria set in PIN_CNF11.SENSE register. Write ‘1’ to clear.

PIN12_R

Field PIN12 reader - Status on whether PIN12 has met criteria set in PIN_CNF12.SENSE register. Write ‘1’ to clear.

PIN12_W

Field PIN12 writer - Status on whether PIN12 has met criteria set in PIN_CNF12.SENSE register. Write ‘1’ to clear.

PIN13_R

Field PIN13 reader - Status on whether PIN13 has met criteria set in PIN_CNF13.SENSE register. Write ‘1’ to clear.

PIN13_W

Field PIN13 writer - Status on whether PIN13 has met criteria set in PIN_CNF13.SENSE register. Write ‘1’ to clear.

PIN14_R

Field PIN14 reader - Status on whether PIN14 has met criteria set in PIN_CNF14.SENSE register. Write ‘1’ to clear.

PIN14_W

Field PIN14 writer - Status on whether PIN14 has met criteria set in PIN_CNF14.SENSE register. Write ‘1’ to clear.

PIN15_R

Field PIN15 reader - Status on whether PIN15 has met criteria set in PIN_CNF15.SENSE register. Write ‘1’ to clear.

PIN15_W

Field PIN15 writer - Status on whether PIN15 has met criteria set in PIN_CNF15.SENSE register. Write ‘1’ to clear.

PIN16_R

Field PIN16 reader - Status on whether PIN16 has met criteria set in PIN_CNF16.SENSE register. Write ‘1’ to clear.

PIN16_W

Field PIN16 writer - Status on whether PIN16 has met criteria set in PIN_CNF16.SENSE register. Write ‘1’ to clear.

PIN17_R

Field PIN17 reader - Status on whether PIN17 has met criteria set in PIN_CNF17.SENSE register. Write ‘1’ to clear.

PIN17_W

Field PIN17 writer - Status on whether PIN17 has met criteria set in PIN_CNF17.SENSE register. Write ‘1’ to clear.

PIN18_R

Field PIN18 reader - Status on whether PIN18 has met criteria set in PIN_CNF18.SENSE register. Write ‘1’ to clear.

PIN18_W

Field PIN18 writer - Status on whether PIN18 has met criteria set in PIN_CNF18.SENSE register. Write ‘1’ to clear.

PIN19_R

Field PIN19 reader - Status on whether PIN19 has met criteria set in PIN_CNF19.SENSE register. Write ‘1’ to clear.

PIN19_W

Field PIN19 writer - Status on whether PIN19 has met criteria set in PIN_CNF19.SENSE register. Write ‘1’ to clear.

PIN20_R

Field PIN20 reader - Status on whether PIN20 has met criteria set in PIN_CNF20.SENSE register. Write ‘1’ to clear.

PIN20_W

Field PIN20 writer - Status on whether PIN20 has met criteria set in PIN_CNF20.SENSE register. Write ‘1’ to clear.

PIN21_R

Field PIN21 reader - Status on whether PIN21 has met criteria set in PIN_CNF21.SENSE register. Write ‘1’ to clear.

PIN21_W

Field PIN21 writer - Status on whether PIN21 has met criteria set in PIN_CNF21.SENSE register. Write ‘1’ to clear.

PIN22_R

Field PIN22 reader - Status on whether PIN22 has met criteria set in PIN_CNF22.SENSE register. Write ‘1’ to clear.

PIN22_W

Field PIN22 writer - Status on whether PIN22 has met criteria set in PIN_CNF22.SENSE register. Write ‘1’ to clear.

PIN23_R

Field PIN23 reader - Status on whether PIN23 has met criteria set in PIN_CNF23.SENSE register. Write ‘1’ to clear.

PIN23_W

Field PIN23 writer - Status on whether PIN23 has met criteria set in PIN_CNF23.SENSE register. Write ‘1’ to clear.

PIN24_R

Field PIN24 reader - Status on whether PIN24 has met criteria set in PIN_CNF24.SENSE register. Write ‘1’ to clear.

PIN24_W

Field PIN24 writer - Status on whether PIN24 has met criteria set in PIN_CNF24.SENSE register. Write ‘1’ to clear.

PIN25_R

Field PIN25 reader - Status on whether PIN25 has met criteria set in PIN_CNF25.SENSE register. Write ‘1’ to clear.

PIN25_W

Field PIN25 writer - Status on whether PIN25 has met criteria set in PIN_CNF25.SENSE register. Write ‘1’ to clear.

PIN26_R

Field PIN26 reader - Status on whether PIN26 has met criteria set in PIN_CNF26.SENSE register. Write ‘1’ to clear.

PIN26_W

Field PIN26 writer - Status on whether PIN26 has met criteria set in PIN_CNF26.SENSE register. Write ‘1’ to clear.

PIN27_R

Field PIN27 reader - Status on whether PIN27 has met criteria set in PIN_CNF27.SENSE register. Write ‘1’ to clear.

PIN27_W

Field PIN27 writer - Status on whether PIN27 has met criteria set in PIN_CNF27.SENSE register. Write ‘1’ to clear.

PIN28_R

Field PIN28 reader - Status on whether PIN28 has met criteria set in PIN_CNF28.SENSE register. Write ‘1’ to clear.

PIN28_W

Field PIN28 writer - Status on whether PIN28 has met criteria set in PIN_CNF28.SENSE register. Write ‘1’ to clear.

PIN29_R

Field PIN29 reader - Status on whether PIN29 has met criteria set in PIN_CNF29.SENSE register. Write ‘1’ to clear.

PIN29_W

Field PIN29 writer - Status on whether PIN29 has met criteria set in PIN_CNF29.SENSE register. Write ‘1’ to clear.

PIN30_R

Field PIN30 reader - Status on whether PIN30 has met criteria set in PIN_CNF30.SENSE register. Write ‘1’ to clear.

PIN30_W

Field PIN30 writer - Status on whether PIN30 has met criteria set in PIN_CNF30.SENSE register. Write ‘1’ to clear.

PIN31_R

Field PIN31 reader - Status on whether PIN31 has met criteria set in PIN_CNF31.SENSE register. Write ‘1’ to clear.

PIN31_W

Field PIN31 writer - Status on whether PIN31 has met criteria set in PIN_CNF31.SENSE register. Write ‘1’ to clear.