Type Alias nrf52840_pac::p0::pin_cnf::DRIVE_R

pub type DRIVE_R = FieldReader<u8, DRIVE_A>;

Field DRIVE reader - Drive configuration

Aliased Type

struct DRIVE_R { /* private fields */ }

Implementations

impl DRIVE_R

pub fn variant(&self) -> DRIVE_A

Get enumerated values variant

pub fn is_s0s1(&self) -> bool

Checks if the value of the field is S0S1

pub fn is_h0s1(&self) -> bool

Checks if the value of the field is H0S1

pub fn is_s0h1(&self) -> bool

Checks if the value of the field is S0H1

pub fn is_h0h1(&self) -> bool

Checks if the value of the field is H0H1

pub fn is_d0s1(&self) -> bool

Checks if the value of the field is D0S1

pub fn is_d0h1(&self) -> bool

Checks if the value of the field is D0H1

pub fn is_s0d1(&self) -> bool

Checks if the value of the field is S0D1

pub fn is_h0d1(&self) -> bool

Checks if the value of the field is H0D1