nrf52840_pac::power
Module pofcon

Power-fail comparator configuration

Structs

POFCON_SPEC

Power-fail comparator configuration

R

Register POFCON reader

W

Register POFCON writer

Enums

POF_A

Enable or disable power failure warning

THRESHOLDVDDH_A

Power-fail comparator threshold setting for high voltage mode (supply connected to VDDH only). This setting does not apply for normal voltage mode (supply connected to both VDD and VDDH).

THRESHOLD_A

Power-fail comparator threshold setting. This setting applies both for normal voltage mode (supply connected to both VDD and VDDH) and high voltage mode (supply connected to VDDH only). Values 0-3 set threshold below 1.7 V and should not be used as brown out detection will be activated before power failure warning on such low voltages.

Type Aliases

POF_R

Field POF reader - Enable or disable power failure warning

POF_W

Field POF writer - Enable or disable power failure warning

THRESHOLDVDDH_R

Field THRESHOLDVDDH reader - Power-fail comparator threshold setting for high voltage mode (supply connected to VDDH only). This setting does not apply for normal voltage mode (supply connected to both VDD and VDDH).

THRESHOLDVDDH_W

Field THRESHOLDVDDH writer - Power-fail comparator threshold setting for high voltage mode (supply connected to VDDH only). This setting does not apply for normal voltage mode (supply connected to both VDD and VDDH).

THRESHOLD_R

Field THRESHOLD reader - Power-fail comparator threshold setting. This setting applies both for normal voltage mode (supply connected to both VDD and VDDH) and high voltage mode (supply connected to VDDH only). Values 0-3 set threshold below 1.7 V and should not be used as brown out detection will be activated before power failure warning on such low voltages.

THRESHOLD_W

Field THRESHOLD writer - Power-fail comparator threshold setting. This setting applies both for normal voltage mode (supply connected to both VDD and VDDH) and high voltage mode (supply connected to VDDH only). Values 0-3 set threshold below 1.7 V and should not be used as brown out detection will be activated before power failure warning on such low voltages.