Type Alias nrf52840_pac::power::pofcon::THRESHOLDVDDH_W

pub type THRESHOLDVDDH_W<'a, const O: u8> = FieldWriterSafe<'a, u32, POFCON_SPEC, u8, THRESHOLDVDDH_A, 4, O>;

Field THRESHOLDVDDH writer - Power-fail comparator threshold setting for high voltage mode (supply connected to VDDH only). This setting does not apply for normal voltage mode (supply connected to both VDD and VDDH).

Aliased Type

struct THRESHOLDVDDH_W<'a, const O: u8> { /* private fields */ }

Implementations

impl<'a, const O: u8> THRESHOLDVDDH_W<'a, O>

pub fn v27(self) -> &'a mut W

Set threshold to 2.7 V

pub fn v28(self) -> &'a mut W

Set threshold to 2.8 V

pub fn v29(self) -> &'a mut W

Set threshold to 2.9 V

pub fn v30(self) -> &'a mut W

Set threshold to 3.0 V

pub fn v31(self) -> &'a mut W

Set threshold to 3.1 V

pub fn v32(self) -> &'a mut W

Set threshold to 3.2 V

pub fn v33(self) -> &'a mut W

Set threshold to 3.3 V

pub fn v34(self) -> &'a mut W

Set threshold to 3.4 V

pub fn v35(self) -> &'a mut W

Set threshold to 3.5 V

pub fn v36(self) -> &'a mut W

Set threshold to 3.6 V

pub fn v37(self) -> &'a mut W

Set threshold to 3.7 V

pub fn v38(self) -> &'a mut W

Set threshold to 3.8 V

pub fn v39(self) -> &'a mut W

Set threshold to 3.9 V

pub fn v40(self) -> &'a mut W

Set threshold to 4.0 V

pub fn v41(self) -> &'a mut W

Set threshold to 4.1 V

pub fn v42(self) -> &'a mut W

Set threshold to 4.2 V