nrf52840_pac::power::pofcon

Type Alias THRESHOLD_R

pub type THRESHOLD_R = FieldReader<u8, THRESHOLD_A>;

Field THRESHOLD reader - Power-fail comparator threshold setting. This setting applies both for normal voltage mode (supply connected to both VDD and VDDH) and high voltage mode (supply connected to VDDH only). Values 0-3 set threshold below 1.7 V and should not be used as brown out detection will be activated before power failure warning on such low voltages.

Aliased Type

struct THRESHOLD_R { /* private fields */ }

Implementations

impl THRESHOLD_R

pub fn variant(&self) -> Option<THRESHOLD_A>

Get enumerated values variant

pub fn is_v17(&self) -> bool

Checks if the value of the field is V17

pub fn is_v18(&self) -> bool

Checks if the value of the field is V18

pub fn is_v19(&self) -> bool

Checks if the value of the field is V19

pub fn is_v20(&self) -> bool

Checks if the value of the field is V20

pub fn is_v21(&self) -> bool

Checks if the value of the field is V21

pub fn is_v22(&self) -> bool

Checks if the value of the field is V22

pub fn is_v23(&self) -> bool

Checks if the value of the field is V23

pub fn is_v24(&self) -> bool

Checks if the value of the field is V24

pub fn is_v25(&self) -> bool

Checks if the value of the field is V25

pub fn is_v26(&self) -> bool

Checks if the value of the field is V26

pub fn is_v27(&self) -> bool

Checks if the value of the field is V27

pub fn is_v28(&self) -> bool

Checks if the value of the field is V28