nrf52840_pac::power::pofcon

Type Alias THRESHOLD_W

pub type THRESHOLD_W<'a, const O: u8> = FieldWriter<'a, u32, POFCON_SPEC, u8, THRESHOLD_A, 4, O>;

Field THRESHOLD writer - Power-fail comparator threshold setting. This setting applies both for normal voltage mode (supply connected to both VDD and VDDH) and high voltage mode (supply connected to VDDH only). Values 0-3 set threshold below 1.7 V and should not be used as brown out detection will be activated before power failure warning on such low voltages.

Aliased Type

struct THRESHOLD_W<'a, const O: u8> { /* private fields */ }

Implementations

impl<'a, const O: u8> THRESHOLD_W<'a, O>

pub fn v17(self) -> &'a mut W

Set threshold to 1.7 V

pub fn v18(self) -> &'a mut W

Set threshold to 1.8 V

pub fn v19(self) -> &'a mut W

Set threshold to 1.9 V

pub fn v20(self) -> &'a mut W

Set threshold to 2.0 V

pub fn v21(self) -> &'a mut W

Set threshold to 2.1 V

pub fn v22(self) -> &'a mut W

Set threshold to 2.2 V

pub fn v23(self) -> &'a mut W

Set threshold to 2.3 V

pub fn v24(self) -> &'a mut W

Set threshold to 2.4 V

pub fn v25(self) -> &'a mut W

Set threshold to 2.5 V

pub fn v26(self) -> &'a mut W

Set threshold to 2.6 V

pub fn v27(self) -> &'a mut W

Set threshold to 2.7 V

pub fn v28(self) -> &'a mut W

Set threshold to 2.8 V