Type Alias nrf52840_pac::power::usbregstatus::VBUSDETECT_R

pub type VBUSDETECT_R = BitReader<VBUSDETECT_A>;

Field VBUSDETECT reader - VBUS input detection status (USBDETECTED and USBREMOVED events are derived from this information)

Aliased Type

struct VBUSDETECT_R { /* private fields */ }

Implementations

impl VBUSDETECT_R

pub fn variant(&self) -> VBUSDETECT_A

Get enumerated values variant

pub fn is_no_vbus(&self) -> bool

Checks if the value of the field is NO_VBUS

pub fn is_vbus_present(&self) -> bool

Checks if the value of the field is VBUS_PRESENT