Type Alias nrf52840_pac::ppi::chg::CH31_R

pub type CH31_R = BitReader<CH31_A>;

Field CH31 reader - Include or exclude channel 31

Aliased Type

struct CH31_R { /* private fields */ }

Implementations

impl CH31_R

pub fn variant(&self) -> CH31_A

Get enumerated values variant

pub fn is_excluded(&self) -> bool

Checks if the value of the field is EXCLUDED

pub fn is_included(&self) -> bool

Checks if the value of the field is INCLUDED