nrf52840_pac::pwm0::decoder

Type Alias LOAD_R

pub type LOAD_R = FieldReader<u8, LOAD_A>;

Field LOAD reader - How a sequence is read from RAM and spread to the compare register

Aliased Type

struct LOAD_R { /* private fields */ }

Implementations

impl LOAD_R

pub fn variant(&self) -> LOAD_A

Get enumerated values variant

pub fn is_common(&self) -> bool

Checks if the value of the field is COMMON

pub fn is_grouped(&self) -> bool

Checks if the value of the field is GROUPED

pub fn is_individual(&self) -> bool

Checks if the value of the field is INDIVIDUAL

pub fn is_wave_form(&self) -> bool

Checks if the value of the field is WAVE_FORM