Module nrf52840_pac::pwm0::seq::refresh

Description cluster[n]: Number of additional PWM periods between samples loaded into compare register

Structs

R

Register REFRESH reader

REFRESH_SPEC

Description cluster[n]: Number of additional PWM periods between samples loaded into compare register

W

Register REFRESH writer

Enums

CNT_A

Number of additional PWM periods between samples loaded into compare register (load every REFRESH.CNT+1 PWM periods)

Type Aliases

CNT_R

Field CNT reader - Number of additional PWM periods between samples loaded into compare register (load every REFRESH.CNT+1 PWM periods)

CNT_W

Field CNT writer - Number of additional PWM periods between samples loaded into compare register (load every REFRESH.CNT+1 PWM periods)