nrf52840_pac::qdec::reportper

Type Alias REPORTPER_R

pub type REPORTPER_R = FieldReader<u8, REPORTPER_A>;

Field REPORTPER reader - Specifies the number of samples to be accumulated in the ACC register before the REPORTRDY and DBLRDY events can be generated

Aliased Type

struct REPORTPER_R { /* private fields */ }

Implementations

impl REPORTPER_R

pub fn variant(&self) -> Option<REPORTPER_A>

Get enumerated values variant

pub fn is_10smpl(&self) -> bool

Checks if the value of the field is _10SMPL

pub fn is_40smpl(&self) -> bool

Checks if the value of the field is _40SMPL

pub fn is_80smpl(&self) -> bool

Checks if the value of the field is _80SMPL

pub fn is_120smpl(&self) -> bool

Checks if the value of the field is _120SMPL

pub fn is_160smpl(&self) -> bool

Checks if the value of the field is _160SMPL

pub fn is_200smpl(&self) -> bool

Checks if the value of the field is _200SMPL

pub fn is_240smpl(&self) -> bool

Checks if the value of the field is _240SMPL

pub fn is_280smpl(&self) -> bool

Checks if the value of the field is _280SMPL

pub fn is_1smpl(&self) -> bool

Checks if the value of the field is _1SMPL