Type Alias nrf52840_pac::qdec::sampleper::SAMPLEPER_R

pub type SAMPLEPER_R = FieldReader<u8, SAMPLEPER_A>;

Field SAMPLEPER reader - Sample period. The SAMPLE register will be updated for every new sample

Aliased Type

struct SAMPLEPER_R { /* private fields */ }

Implementations

impl SAMPLEPER_R

pub fn variant(&self) -> Option<SAMPLEPER_A>

Get enumerated values variant

pub fn is_128us(&self) -> bool

Checks if the value of the field is _128US

pub fn is_256us(&self) -> bool

Checks if the value of the field is _256US

pub fn is_512us(&self) -> bool

Checks if the value of the field is _512US

pub fn is_1024us(&self) -> bool

Checks if the value of the field is _1024US

pub fn is_2048us(&self) -> bool

Checks if the value of the field is _2048US

pub fn is_4096us(&self) -> bool

Checks if the value of the field is _4096US

pub fn is_8192us(&self) -> bool

Checks if the value of the field is _8192US

pub fn is_16384us(&self) -> bool

Checks if the value of the field is _16384US

pub fn is_32ms(&self) -> bool

Checks if the value of the field is _32MS

pub fn is_65ms(&self) -> bool

Checks if the value of the field is _65MS

pub fn is_131ms(&self) -> bool

Checks if the value of the field is _131MS