Module nrf52840_pac::qspi

External flash interface

Re-exports

pub use read::READ;

pub use write::WRITE;

pub use erase::ERASE;

pub use psel::PSEL;

Modules

addrconf

Extended address configuration.

cinstrconf

Custom instruction configuration register.

cinstrdat0

Custom instruction data register 0.

cinstrdat1

Custom instruction data register 1.

dpmdur

Set the duration required to enter/exit deep power-down mode (DPM).

enable

Enable QSPI peripheral and acquire the pins selected in PSELn registers

erase

Cluster Unspecified

events_ready

QSPI peripheral is ready. This event will be generated as a response to any QSPI task.

ifconfig0

Interface configuration.

ifconfig1

Interface configuration.

iftiming

SPI interface timing.

inten

Enable or disable interrupt

intenclr

Disable interrupt

intenset

Enable interrupt

psel

Cluster Unspecified

read

Cluster Unspecified

status

Status register.

tasks_activate

Activate QSPI interface

tasks_deactivate

Deactivate QSPI interface

tasks_erasestart

Start external flash memory erase operation

tasks_readstart

Start transfer from external flash memory to internal RAM

tasks_writestart

Start transfer from internal RAM to external flash memory

write

Cluster Unspecified

xipoffset

Address offset into the external memory for Execute in Place operation.

Structs

RegisterBlock

Register block

Type Aliases

ADDRCONF

ADDRCONF (rw) register accessor: an alias for Reg<ADDRCONF_SPEC>

CINSTRCONF

CINSTRCONF (rw) register accessor: an alias for Reg<CINSTRCONF_SPEC>

CINSTRDAT0

CINSTRDAT0 (rw) register accessor: an alias for Reg<CINSTRDAT0_SPEC>

CINSTRDAT1

CINSTRDAT1 (rw) register accessor: an alias for Reg<CINSTRDAT1_SPEC>

DPMDUR

DPMDUR (rw) register accessor: an alias for Reg<DPMDUR_SPEC>

ENABLE

ENABLE (rw) register accessor: an alias for Reg<ENABLE_SPEC>

EVENTS_READY

EVENTS_READY (rw) register accessor: an alias for Reg<EVENTS_READY_SPEC>

IFCONFIG0

IFCONFIG0 (rw) register accessor: an alias for Reg<IFCONFIG0_SPEC>

IFCONFIG1

IFCONFIG1 (rw) register accessor: an alias for Reg<IFCONFIG1_SPEC>

IFTIMING

IFTIMING (rw) register accessor: an alias for Reg<IFTIMING_SPEC>

INTEN

INTEN (rw) register accessor: an alias for Reg<INTEN_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

STATUS

STATUS (r) register accessor: an alias for Reg<STATUS_SPEC>

TASKS_ACTIVATE

TASKS_ACTIVATE (w) register accessor: an alias for Reg<TASKS_ACTIVATE_SPEC>

TASKS_DEACTIVATE

TASKS_DEACTIVATE (w) register accessor: an alias for Reg<TASKS_DEACTIVATE_SPEC>

TASKS_ERASESTART

TASKS_ERASESTART (w) register accessor: an alias for Reg<TASKS_ERASESTART_SPEC>

TASKS_READSTART

TASKS_READSTART (w) register accessor: an alias for Reg<TASKS_READSTART_SPEC>

TASKS_WRITESTART

TASKS_WRITESTART (w) register accessor: an alias for Reg<TASKS_WRITESTART_SPEC>

XIPOFFSET

XIPOFFSET (rw) register accessor: an alias for Reg<XIPOFFSET_SPEC>