nrf52840_pac::qspi::addrconf

Type Alias MODE_R

pub type MODE_R = FieldReader<u8, MODE_A>;

Field MODE reader - Extended addressing mode.

Aliased Type

struct MODE_R { /* private fields */ }

Implementations

impl MODE_R

pub fn variant(&self) -> MODE_A

Get enumerated values variant

pub fn is_no_instr(&self) -> bool

Checks if the value of the field is NO_INSTR

pub fn is_opcode(&self) -> bool

Checks if the value of the field is OPCODE

pub fn is_op_byte0(&self) -> bool

Checks if the value of the field is OP_BYTE0

pub fn is_all(&self) -> bool

Checks if the value of the field is ALL