nrf52840_pac::qspi::cinstrconf

Type Alias LENGTH_R

pub type LENGTH_R = FieldReader<u8, LENGTH_A>;

Field LENGTH reader - Length of custom instruction in number of bytes.

Aliased Type

struct LENGTH_R { /* private fields */ }

Implementations

impl LENGTH_R

pub fn variant(&self) -> Option<LENGTH_A>

Get enumerated values variant

pub fn is_1b(&self) -> bool

Checks if the value of the field is _1B

pub fn is_2b(&self) -> bool

Checks if the value of the field is _2B

pub fn is_3b(&self) -> bool

Checks if the value of the field is _3B

pub fn is_4b(&self) -> bool

Checks if the value of the field is _4B

pub fn is_5b(&self) -> bool

Checks if the value of the field is _5B

pub fn is_6b(&self) -> bool

Checks if the value of the field is _6B

pub fn is_7b(&self) -> bool

Checks if the value of the field is _7B

pub fn is_8b(&self) -> bool

Checks if the value of the field is _8B

pub fn is_9b(&self) -> bool

Checks if the value of the field is _9B