nrf52840_pac::qspi
Module ifconfig0

Interface configuration.

Structs

IFCONFIG0_SPEC

Interface configuration.

R

Register IFCONFIG0 reader

W

Register IFCONFIG0 writer

Enums

ADDRMODE_A

Addressing mode.

DPMENABLE_A

Enable deep power-down mode (DPM) feature.

PPSIZE_A

Page size for commands PP, PP2O, PP4O and PP4IO.

READOC_A

Configure number of data lines and opcode used for reading.

WRITEOC_A

Configure number of data lines and opcode used for writing.

Type Aliases

ADDRMODE_R

Field ADDRMODE reader - Addressing mode.

ADDRMODE_W

Field ADDRMODE writer - Addressing mode.

DPMENABLE_R

Field DPMENABLE reader - Enable deep power-down mode (DPM) feature.

DPMENABLE_W

Field DPMENABLE writer - Enable deep power-down mode (DPM) feature.

PPSIZE_R

Field PPSIZE reader - Page size for commands PP, PP2O, PP4O and PP4IO.

PPSIZE_W

Field PPSIZE writer - Page size for commands PP, PP2O, PP4O and PP4IO.

READOC_R

Field READOC reader - Configure number of data lines and opcode used for reading.

READOC_W

Field READOC writer - Configure number of data lines and opcode used for reading.

WRITEOC_R

Field WRITEOC reader - Configure number of data lines and opcode used for writing.

WRITEOC_W

Field WRITEOC writer - Configure number of data lines and opcode used for writing.