Module nrf52840_pac::qspi::ifconfig1

Interface configuration.

Structs

IFCONFIG1_SPEC

Interface configuration.

R

Register IFCONFIG1 reader

W

Register IFCONFIG1 writer

Enums

DPMEN_A

Enter/exit deep power-down mode (DPM) for external flash memory.

SPIMODE_A

Select SPI mode.

Type Aliases

DPMEN_R

Field DPMEN reader - Enter/exit deep power-down mode (DPM) for external flash memory.

DPMEN_W

Field DPMEN writer - Enter/exit deep power-down mode (DPM) for external flash memory.

SCKDELAY_R

Field SCKDELAY reader - Minimum amount of time that the CSN pin must stay high before it can go low again. Value is specified in number of 16 MHz periods (62.5 ns).

SCKDELAY_W

Field SCKDELAY writer - Minimum amount of time that the CSN pin must stay high before it can go low again. Value is specified in number of 16 MHz periods (62.5 ns).

SCKFREQ_R

Field SCKFREQ reader - SCK frequency is given as 32 MHz / (SCKFREQ + 1).

SCKFREQ_W

Field SCKFREQ writer - SCK frequency is given as 32 MHz / (SCKFREQ + 1).

SPIMODE_R

Field SPIMODE reader - Select SPI mode.

SPIMODE_W

Field SPIMODE writer - Select SPI mode.