nrf52840_pac::qspi::ifconfig1
Type Alias SPIMODE_R

pub type SPIMODE_R = BitReader<SPIMODE_A>;

Field SPIMODE reader - Select SPI mode.

Aliased Type

struct SPIMODE_R { /* private fields */ }

Implementations

impl SPIMODE_R

pub fn variant(&self) -> SPIMODE_A

Get enumerated values variant

pub fn is_mode0(&self) -> bool

Checks if the value of the field is MODE0

pub fn is_mode3(&self) -> bool

Checks if the value of the field is MODE3