Struct nrf52840_pac::radio::dacnf::R

pub struct R(/* private fields */);

Register DACNF reader

Implementations

impl R

pub fn ena0(&self) -> ENA0_R

Bit 0 - Enable or disable device address matching using device address 0

pub fn ena1(&self) -> ENA1_R

Bit 1 - Enable or disable device address matching using device address 1

pub fn ena2(&self) -> ENA2_R

Bit 2 - Enable or disable device address matching using device address 2

pub fn ena3(&self) -> ENA3_R

Bit 3 - Enable or disable device address matching using device address 3

pub fn ena4(&self) -> ENA4_R

Bit 4 - Enable or disable device address matching using device address 4

pub fn ena5(&self) -> ENA5_R

Bit 5 - Enable or disable device address matching using device address 5

pub fn ena6(&self) -> ENA6_R

Bit 6 - Enable or disable device address matching using device address 6

pub fn ena7(&self) -> ENA7_R

Bit 7 - Enable or disable device address matching using device address 7

pub fn txadd0(&self) -> TXADD0_R

Bit 8 - TxAdd for device address 0

pub fn txadd1(&self) -> TXADD1_R

Bit 9 - TxAdd for device address 1

pub fn txadd2(&self) -> TXADD2_R

Bit 10 - TxAdd for device address 2

pub fn txadd3(&self) -> TXADD3_R

Bit 11 - TxAdd for device address 3

pub fn txadd4(&self) -> TXADD4_R

Bit 12 - TxAdd for device address 4

pub fn txadd5(&self) -> TXADD5_R

Bit 13 - TxAdd for device address 5

pub fn txadd6(&self) -> TXADD6_R

Bit 14 - TxAdd for device address 6

pub fn txadd7(&self) -> TXADD7_R

Bit 15 - TxAdd for device address 7