nrf52840_pac::radio::mode

Type Alias MODE_R

pub type MODE_R = FieldReader<u8, MODE_A>;

Field MODE reader - Radio data rate and modulation setting. The radio supports frequency-shift keying (FSK) modulation.

Aliased Type

struct MODE_R { /* private fields */ }

Implementations

impl MODE_R

pub fn variant(&self) -> Option<MODE_A>

Get enumerated values variant

pub fn is_nrf_1mbit(&self) -> bool

Checks if the value of the field is NRF_1MBIT

pub fn is_nrf_2mbit(&self) -> bool

Checks if the value of the field is NRF_2MBIT

pub fn is_ble_1mbit(&self) -> bool

Checks if the value of the field is BLE_1MBIT

pub fn is_ble_2mbit(&self) -> bool

Checks if the value of the field is BLE_2MBIT

pub fn is_ble_lr125kbit(&self) -> bool

Checks if the value of the field is BLE_LR125KBIT

pub fn is_ble_lr500kbit(&self) -> bool

Checks if the value of the field is BLE_LR500KBIT

pub fn is_ieee802154_250kbit(&self) -> bool

Checks if the value of the field is IEEE802154_250KBIT