nrf52840_pac::radio

Module pcnf1

Packet configuration register 1

Structs

PCNF1_SPEC

Packet configuration register 1

R

Register PCNF1 reader

W

Register PCNF1 writer

Enums

ENDIAN_A

On air endianness of packet, this applies to the S0, LENGTH, S1 and the PAYLOAD fields.

WHITEEN_A

Enable or disable packet whitening

Type Aliases

BALEN_R

Field BALEN reader - Base address length in number of bytes

BALEN_W

Field BALEN writer - Base address length in number of bytes

ENDIAN_R

Field ENDIAN reader - On air endianness of packet, this applies to the S0, LENGTH, S1 and the PAYLOAD fields.

ENDIAN_W

Field ENDIAN writer - On air endianness of packet, this applies to the S0, LENGTH, S1 and the PAYLOAD fields.

MAXLEN_R

Field MAXLEN reader - Maximum length of packet payload. If the packet payload is larger than MAXLEN, the radio will truncate the payload to MAXLEN.

MAXLEN_W

Field MAXLEN writer - Maximum length of packet payload. If the packet payload is larger than MAXLEN, the radio will truncate the payload to MAXLEN.

STATLEN_R

Field STATLEN reader - Static length in number of bytes

STATLEN_W

Field STATLEN writer - Static length in number of bytes

WHITEEN_R

Field WHITEEN reader - Enable or disable packet whitening

WHITEEN_W

Field WHITEEN writer - Enable or disable packet whitening