nrf52840_pac::radio::state

Type Alias STATE_R

pub type STATE_R = FieldReader<u8, STATE_A>;

Field STATE reader - Current radio state

Aliased Type

struct STATE_R { /* private fields */ }

Implementations

impl STATE_R

pub fn variant(&self) -> Option<STATE_A>

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_rx_ru(&self) -> bool

Checks if the value of the field is RX_RU

pub fn is_rx_idle(&self) -> bool

Checks if the value of the field is RX_IDLE

pub fn is_rx(&self) -> bool

Checks if the value of the field is RX

pub fn is_rx_disable(&self) -> bool

Checks if the value of the field is RX_DISABLE

pub fn is_tx_ru(&self) -> bool

Checks if the value of the field is TX_RU

pub fn is_tx_idle(&self) -> bool

Checks if the value of the field is TX_IDLE

pub fn is_tx(&self) -> bool

Checks if the value of the field is TX

pub fn is_tx_disable(&self) -> bool

Checks if the value of the field is TX_DISABLE