nrf52840_pac::radio

Struct RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub tasks_txen: TASKS_TXEN,
    pub tasks_rxen: TASKS_RXEN,
    pub tasks_start: TASKS_START,
    pub tasks_stop: TASKS_STOP,
    pub tasks_disable: TASKS_DISABLE,
    pub tasks_rssistart: TASKS_RSSISTART,
    pub tasks_rssistop: TASKS_RSSISTOP,
    pub tasks_bcstart: TASKS_BCSTART,
    pub tasks_bcstop: TASKS_BCSTOP,
    pub tasks_edstart: TASKS_EDSTART,
    pub tasks_edstop: TASKS_EDSTOP,
    pub tasks_ccastart: TASKS_CCASTART,
    pub tasks_ccastop: TASKS_CCASTOP,
    pub events_ready: EVENTS_READY,
    pub events_address: EVENTS_ADDRESS,
    pub events_payload: EVENTS_PAYLOAD,
    pub events_end: EVENTS_END,
    pub events_disabled: EVENTS_DISABLED,
    pub events_devmatch: EVENTS_DEVMATCH,
    pub events_devmiss: EVENTS_DEVMISS,
    pub events_rssiend: EVENTS_RSSIEND,
    pub events_bcmatch: EVENTS_BCMATCH,
    pub events_crcok: EVENTS_CRCOK,
    pub events_crcerror: EVENTS_CRCERROR,
    pub events_framestart: EVENTS_FRAMESTART,
    pub events_edend: EVENTS_EDEND,
    pub events_edstopped: EVENTS_EDSTOPPED,
    pub events_ccaidle: EVENTS_CCAIDLE,
    pub events_ccabusy: EVENTS_CCABUSY,
    pub events_ccastopped: EVENTS_CCASTOPPED,
    pub events_rateboost: EVENTS_RATEBOOST,
    pub events_txready: EVENTS_TXREADY,
    pub events_rxready: EVENTS_RXREADY,
    pub events_mhrmatch: EVENTS_MHRMATCH,
    pub events_phyend: EVENTS_PHYEND,
    pub shorts: SHORTS,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub crcstatus: CRCSTATUS,
    pub rxmatch: RXMATCH,
    pub rxcrc: RXCRC,
    pub dai: DAI,
    pub pdustat: PDUSTAT,
    pub packetptr: PACKETPTR,
    pub frequency: FREQUENCY,
    pub txpower: TXPOWER,
    pub mode: MODE,
    pub pcnf0: PCNF0,
    pub pcnf1: PCNF1,
    pub base0: BASE0,
    pub base1: BASE1,
    pub prefix0: PREFIX0,
    pub prefix1: PREFIX1,
    pub txaddress: TXADDRESS,
    pub rxaddresses: RXADDRESSES,
    pub crccnf: CRCCNF,
    pub crcpoly: CRCPOLY,
    pub crcinit: CRCINIT,
    pub tifs: TIFS,
    pub rssisample: RSSISAMPLE,
    pub state: STATE,
    pub datawhiteiv: DATAWHITEIV,
    pub bcc: BCC,
    pub dab: [DAB; 8],
    pub dap: [DAP; 8],
    pub dacnf: DACNF,
    pub mhrmatchconf: MHRMATCHCONF,
    pub mhrmatchmas: MHRMATCHMAS,
    pub modecnf0: MODECNF0,
    pub sfd: SFD,
    pub edcnt: EDCNT,
    pub edsample: EDSAMPLE,
    pub ccactrl: CCACTRL,
    pub power: POWER,
    /* private fields */

}

Register block

Fields
tasks_txen: TASKS_TXEN

0x00 - Enable RADIO in TX mode

tasks_rxen: TASKS_RXEN

0x04 - Enable RADIO in RX mode

tasks_start: TASKS_START

0x08 - Start RADIO

tasks_stop: TASKS_STOP

0x0c - Stop RADIO

tasks_disable: TASKS_DISABLE

0x10 - Disable RADIO

tasks_rssistart: TASKS_RSSISTART

0x14 - Start the RSSI and take one single sample of the receive signal strength

tasks_rssistop: TASKS_RSSISTOP

0x18 - Stop the RSSI measurement

tasks_bcstart: TASKS_BCSTART

0x1c - Start the bit counter

tasks_bcstop: TASKS_BCSTOP

0x20 - Stop the bit counter

tasks_edstart: TASKS_EDSTART

0x24 - Start the energy detect measurement used in IEEE 802.15.4 mode

tasks_edstop: TASKS_EDSTOP

0x28 - Stop the energy detect measurement

tasks_ccastart: TASKS_CCASTART

0x2c - Start the clear channel assessment used in IEEE 802.15.4 mode

tasks_ccastop: TASKS_CCASTOP

0x30 - Stop the clear channel assessment

events_ready: EVENTS_READY

0x100 - RADIO has ramped up and is ready to be started

events_address: EVENTS_ADDRESS

0x104 - Address sent or received

events_payload: EVENTS_PAYLOAD

0x108 - Packet payload sent or received

events_end: EVENTS_END

0x10c - Packet sent or received

events_disabled: EVENTS_DISABLED

0x110 - RADIO has been disabled

events_devmatch: EVENTS_DEVMATCH

0x114 - A device address match occurred on the last received packet

events_devmiss: EVENTS_DEVMISS

0x118 - No device address match occurred on the last received packet

events_rssiend: EVENTS_RSSIEND

0x11c - Sampling of receive signal strength complete

events_bcmatch: EVENTS_BCMATCH

0x128 - Bit counter reached bit count value

events_crcok: EVENTS_CRCOK

0x130 - Packet received with CRC ok

events_crcerror: EVENTS_CRCERROR

0x134 - Packet received with CRC error

events_framestart: EVENTS_FRAMESTART

0x138 - IEEE 802.15.4 length field received

events_edend: EVENTS_EDEND

0x13c - Sampling of energy detection complete. A new ED sample is ready for readout from the RADIO.EDSAMPLE register.

events_edstopped: EVENTS_EDSTOPPED

0x140 - The sampling of energy detection has stopped

events_ccaidle: EVENTS_CCAIDLE

0x144 - Wireless medium in idle - clear to send

events_ccabusy: EVENTS_CCABUSY

0x148 - Wireless medium busy - do not send

events_ccastopped: EVENTS_CCASTOPPED

0x14c - The CCA has stopped

events_rateboost: EVENTS_RATEBOOST

0x150 - Ble_LR CI field received, receive mode is changed from Ble_LR125Kbit to Ble_LR500Kbit.

events_txready: EVENTS_TXREADY

0x154 - RADIO has ramped up and is ready to be started TX path

events_rxready: EVENTS_RXREADY

0x158 - RADIO has ramped up and is ready to be started RX path

events_mhrmatch: EVENTS_MHRMATCH

0x15c - MAC header match found

events_phyend: EVENTS_PHYEND

0x16c - Generated in Ble_LR125Kbit, Ble_LR500Kbit and BleIeee802154_250Kbit modes when last bit is sent on air.

shorts: SHORTS

0x200 - Shortcut register

intenset: INTENSET

0x304 - Enable interrupt

intenclr: INTENCLR

0x308 - Disable interrupt

crcstatus: CRCSTATUS

0x400 - CRC status

rxmatch: RXMATCH

0x408 - Received address

rxcrc: RXCRC

0x40c - CRC field of previously received packet

dai: DAI

0x410 - Device address match index

pdustat: PDUSTAT

0x414 - Payload status

packetptr: PACKETPTR

0x504 - Packet pointer

frequency: FREQUENCY

0x508 - Frequency

txpower: TXPOWER

0x50c - Output power

mode: MODE

0x510 - Data rate and modulation

pcnf0: PCNF0

0x514 - Packet configuration register 0

pcnf1: PCNF1

0x518 - Packet configuration register 1

base0: BASE0

0x51c - Base address 0

base1: BASE1

0x520 - Base address 1

prefix0: PREFIX0

0x524 - Prefixes bytes for logical addresses 0-3

prefix1: PREFIX1

0x528 - Prefixes bytes for logical addresses 4-7

txaddress: TXADDRESS

0x52c - Transmit address select

rxaddresses: RXADDRESSES

0x530 - Receive address select

crccnf: CRCCNF

0x534 - CRC configuration

crcpoly: CRCPOLY

0x538 - CRC polynomial

crcinit: CRCINIT

0x53c - CRC initial value

tifs: TIFS

0x544 - Interframe spacing in us

rssisample: RSSISAMPLE

0x548 - RSSI sample

state: STATE

0x550 - Current radio state

datawhiteiv: DATAWHITEIV

0x554 - Data whitening initial value

bcc: BCC

0x560 - Bit counter compare

dab: [DAB; 8]

0x600..0x620 - Description collection[n]: Device address base segment n

dap: [DAP; 8]

0x620..0x640 - Description collection[n]: Device address prefix n

dacnf: DACNF

0x640 - Device address match configuration

mhrmatchconf: MHRMATCHCONF

0x644 - Search pattern configuration

mhrmatchmas: MHRMATCHMAS

0x648 - Pattern mask

modecnf0: MODECNF0

0x650 - Radio mode configuration register 0

sfd: SFD

0x660 - IEEE 802.15.4 start of frame delimiter

edcnt: EDCNT

0x664 - IEEE 802.15.4 energy detect loop count

edsample: EDSAMPLE

0x668 - IEEE 802.15.4 energy detect level

ccactrl: CCACTRL

0x66c - IEEE 802.15.4 clear channel assessment control

power: POWER

0xffc - Peripheral power control