nrf52840_pac::radio::txpower

Type Alias TXPOWER_R

pub type TXPOWER_R = FieldReader<u8, TXPOWER_A>;

Field TXPOWER reader - RADIO output power

Aliased Type

struct TXPOWER_R { /* private fields */ }

Implementations

impl TXPOWER_R

pub fn variant(&self) -> Option<TXPOWER_A>

Get enumerated values variant

pub fn is_pos8d_bm(&self) -> bool

Checks if the value of the field is POS8D_BM

pub fn is_pos7d_bm(&self) -> bool

Checks if the value of the field is POS7D_BM

pub fn is_pos6d_bm(&self) -> bool

Checks if the value of the field is POS6D_BM

pub fn is_pos5d_bm(&self) -> bool

Checks if the value of the field is POS5D_BM

pub fn is_pos4d_bm(&self) -> bool

Checks if the value of the field is POS4D_BM

pub fn is_pos3d_bm(&self) -> bool

Checks if the value of the field is POS3D_BM

pub fn is_pos2d_bm(&self) -> bool

Checks if the value of the field is POS2D_BM

pub fn is_0d_bm(&self) -> bool

Checks if the value of the field is _0D_BM

pub fn is_neg4d_bm(&self) -> bool

Checks if the value of the field is NEG4D_BM

pub fn is_neg8d_bm(&self) -> bool

Checks if the value of the field is NEG8D_BM

pub fn is_neg12d_bm(&self) -> bool

Checks if the value of the field is NEG12D_BM

pub fn is_neg16d_bm(&self) -> bool

Checks if the value of the field is NEG16D_BM

pub fn is_neg20d_bm(&self) -> bool

Checks if the value of the field is NEG20D_BM

pub fn is_neg30d_bm(&self) -> bool

Checks if the value of the field is NEG30D_BM

pub fn is_neg40d_bm(&self) -> bool

Checks if the value of the field is NEG40D_BM