Module nrf52840_pac::saadc

Successive approximation register (SAR) analog-to-digital converter

Re-exports

pub use events_ch::EVENTS_CH;

pub use ch::CH;

pub use result::RESULT;

Modules

ch

Cluster Unspecified

enable

Enable or disable SAADC

events_calibratedone

Calibration is complete

events_ch

Cluster Unspecified

events_done

A conversion task has been completed. Depending on the configuration, multiple conversions might be needed for a result to be transferred to RAM.

events_end

The SAADC has filled up the result buffer

events_resultdone

Result ready for transfer to RAM

events_started

The SAADC has started

events_stopped

The SAADC has stopped

inten

Enable or disable interrupt

intenclr

Disable interrupt

intenset

Enable interrupt

oversample

Oversampling configuration. The RESOLUTION is applied before averaging, thus for high OVERSAMPLE a higher RESOLUTION should be used.

resolution

Resolution configuration

result

Cluster RESULT EasyDMA channel

samplerate

Controls normal or continuous sample rate

status

Status

tasks_calibrateoffset

Starts offset auto-calibration

tasks_sample

Takes one SAADC sample

tasks_start

Starts the SAADC and prepares the result buffer in RAM

tasks_stop

Stops the SAADC and terminates all on-going conversions

Structs

RegisterBlock

Register block

Type Aliases

ENABLE

ENABLE (rw) register accessor: an alias for Reg<ENABLE_SPEC>

EVENTS_CALIBRATEDONE

EVENTS_CALIBRATEDONE (rw) register accessor: an alias for Reg<EVENTS_CALIBRATEDONE_SPEC>

EVENTS_DONE

EVENTS_DONE (rw) register accessor: an alias for Reg<EVENTS_DONE_SPEC>

EVENTS_END

EVENTS_END (rw) register accessor: an alias for Reg<EVENTS_END_SPEC>

EVENTS_RESULTDONE

EVENTS_RESULTDONE (rw) register accessor: an alias for Reg<EVENTS_RESULTDONE_SPEC>

EVENTS_STARTED

EVENTS_STARTED (rw) register accessor: an alias for Reg<EVENTS_STARTED_SPEC>

EVENTS_STOPPED

EVENTS_STOPPED (rw) register accessor: an alias for Reg<EVENTS_STOPPED_SPEC>

INTEN

INTEN (rw) register accessor: an alias for Reg<INTEN_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

OVERSAMPLE

OVERSAMPLE (rw) register accessor: an alias for Reg<OVERSAMPLE_SPEC>

RESOLUTION

RESOLUTION (rw) register accessor: an alias for Reg<RESOLUTION_SPEC>

SAMPLERATE

SAMPLERATE (rw) register accessor: an alias for Reg<SAMPLERATE_SPEC>

STATUS

STATUS (r) register accessor: an alias for Reg<STATUS_SPEC>

TASKS_CALIBRATEOFFSET

TASKS_CALIBRATEOFFSET (w) register accessor: an alias for Reg<TASKS_CALIBRATEOFFSET_SPEC>

TASKS_SAMPLE

TASKS_SAMPLE (w) register accessor: an alias for Reg<TASKS_SAMPLE_SPEC>

TASKS_START

TASKS_START (w) register accessor: an alias for Reg<TASKS_START_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>