nrf52840_pac::saadc::ch::config
Type Alias GAIN_R

pub type GAIN_R = FieldReader<u8, GAIN_A>;

Field GAIN reader - Gain control

Aliased Type

struct GAIN_R { /* private fields */ }

Implementations

impl GAIN_R

pub fn variant(&self) -> GAIN_A

Get enumerated values variant

pub fn is_gain1_6(&self) -> bool

Checks if the value of the field is GAIN1_6

pub fn is_gain1_5(&self) -> bool

Checks if the value of the field is GAIN1_5

pub fn is_gain1_4(&self) -> bool

Checks if the value of the field is GAIN1_4

pub fn is_gain1_3(&self) -> bool

Checks if the value of the field is GAIN1_3

pub fn is_gain1_2(&self) -> bool

Checks if the value of the field is GAIN1_2

pub fn is_gain1(&self) -> bool

Checks if the value of the field is GAIN1

pub fn is_gain2(&self) -> bool

Checks if the value of the field is GAIN2

pub fn is_gain4(&self) -> bool

Checks if the value of the field is GAIN4