nrf52840_pac::saadc::ch::config
Type Alias RESP_R

pub type RESP_R = FieldReader<u8, RESP_A>;

Field RESP reader - Positive channel resistor control

Aliased Type

struct RESP_R { /* private fields */ }

Implementations

impl RESP_R

pub fn variant(&self) -> RESP_A

Get enumerated values variant

pub fn is_bypass(&self) -> bool

Checks if the value of the field is BYPASS

pub fn is_pulldown(&self) -> bool

Checks if the value of the field is PULLDOWN

pub fn is_pullup(&self) -> bool

Checks if the value of the field is PULLUP

pub fn is_vdd1_2(&self) -> bool

Checks if the value of the field is VDD1_2