nrf52840_pac::saadc::ch::pseln

Type Alias PSELN_R

pub type PSELN_R = FieldReader<u8, PSELN_A>;

Field PSELN reader - Analog negative input, enables differential channel

Aliased Type

struct PSELN_R { /* private fields */ }

Implementations

impl PSELN_R

pub fn variant(&self) -> Option<PSELN_A>

Get enumerated values variant

pub fn is_nc(&self) -> bool

Checks if the value of the field is NC

pub fn is_analog_input0(&self) -> bool

Checks if the value of the field is ANALOG_INPUT0

pub fn is_analog_input1(&self) -> bool

Checks if the value of the field is ANALOG_INPUT1

pub fn is_analog_input2(&self) -> bool

Checks if the value of the field is ANALOG_INPUT2

pub fn is_analog_input3(&self) -> bool

Checks if the value of the field is ANALOG_INPUT3

pub fn is_analog_input4(&self) -> bool

Checks if the value of the field is ANALOG_INPUT4

pub fn is_analog_input5(&self) -> bool

Checks if the value of the field is ANALOG_INPUT5

pub fn is_analog_input6(&self) -> bool

Checks if the value of the field is ANALOG_INPUT6

pub fn is_analog_input7(&self) -> bool

Checks if the value of the field is ANALOG_INPUT7

pub fn is_vdd(&self) -> bool

Checks if the value of the field is VDD

pub fn is_vddhdiv5(&self) -> bool

Checks if the value of the field is VDDHDIV5