nrf52840_pac::saadc

Struct RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub tasks_start: TASKS_START,
    pub tasks_sample: TASKS_SAMPLE,
    pub tasks_stop: TASKS_STOP,
    pub tasks_calibrateoffset: TASKS_CALIBRATEOFFSET,
    pub events_started: EVENTS_STARTED,
    pub events_end: EVENTS_END,
    pub events_done: EVENTS_DONE,
    pub events_resultdone: EVENTS_RESULTDONE,
    pub events_calibratedone: EVENTS_CALIBRATEDONE,
    pub events_stopped: EVENTS_STOPPED,
    pub events_ch: [EVENTS_CH; 8],
    pub inten: INTEN,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub status: STATUS,
    pub enable: ENABLE,
    pub ch: [CH; 8],
    pub resolution: RESOLUTION,
    pub oversample: OVERSAMPLE,
    pub samplerate: SAMPLERATE,
    pub result: RESULT,
    /* private fields */

}

Register block

Fields
tasks_start: TASKS_START

0x00 - Starts the SAADC and prepares the result buffer in RAM

tasks_sample: TASKS_SAMPLE

0x04 - Takes one SAADC sample

tasks_stop: TASKS_STOP

0x08 - Stops the SAADC and terminates all on-going conversions

tasks_calibrateoffset: TASKS_CALIBRATEOFFSET

0x0c - Starts offset auto-calibration

events_started: EVENTS_STARTED

0x100 - The SAADC has started

events_end: EVENTS_END

0x104 - The SAADC has filled up the result buffer

events_done: EVENTS_DONE

0x108 - A conversion task has been completed. Depending on the configuration, multiple conversions might be needed for a result to be transferred to RAM.

events_resultdone: EVENTS_RESULTDONE

0x10c - Result ready for transfer to RAM

events_calibratedone: EVENTS_CALIBRATEDONE

0x110 - Calibration is complete

events_stopped: EVENTS_STOPPED

0x114 - The SAADC has stopped

events_ch: [EVENTS_CH; 8]

0x118..0x158 - Unspecified

inten: INTEN

0x300 - Enable or disable interrupt

intenset: INTENSET

0x304 - Enable interrupt

intenclr: INTENCLR

0x308 - Disable interrupt

status: STATUS

0x400 - Status

enable: ENABLE

0x500 - Enable or disable SAADC

ch: [CH; 8]

0x510..0x590 - Unspecified

resolution: RESOLUTION

0x5f0 - Resolution configuration

oversample: OVERSAMPLE

0x5f4 - Oversampling configuration. The RESOLUTION is applied before averaging, thus for high OVERSAMPLE a higher RESOLUTION should be used.

samplerate: SAMPLERATE

0x5f8 - Controls normal or continuous sample rate

result: RESULT

0x62c..0x638 - RESULT EasyDMA channel