Type Alias nrf52840_pac::spi0::frequency::FREQUENCY_R

pub type FREQUENCY_R = FieldReader<u32, FREQUENCY_A>;

Field FREQUENCY reader - SPI master data rate

Aliased Type

struct FREQUENCY_R { /* private fields */ }

Implementations

impl FREQUENCY_R

pub fn variant(&self) -> Option<FREQUENCY_A>

Get enumerated values variant

pub fn is_k125(&self) -> bool

Checks if the value of the field is K125

pub fn is_k250(&self) -> bool

Checks if the value of the field is K250

pub fn is_k500(&self) -> bool

Checks if the value of the field is K500

pub fn is_m1(&self) -> bool

Checks if the value of the field is M1

pub fn is_m2(&self) -> bool

Checks if the value of the field is M2

pub fn is_m4(&self) -> bool

Checks if the value of the field is M4

pub fn is_m8(&self) -> bool

Checks if the value of the field is M8