nrf52840_pac::spim0::iftiming
Module rxdelay

Sample delay for input serial data on MISO

Structs

R

Register RXDELAY reader

RXDELAY_SPEC

Sample delay for input serial data on MISO

W

Register RXDELAY writer

Type Aliases

RXDELAY_R

Field RXDELAY reader - Sample delay for input serial data on MISO. The value specifies the number of 64 MHz clock cycles (15.625 ns) delay from the the sampling edge of SCK (leading edge for CONFIG.CPHA = 0, trailing edge for CONFIG.CPHA = 1) until the input serial data is sampled. As en example, if RXDELAY = 0 and CONFIG.CPHA = 0, the input serial data is sampled on the rising edge of SCK.

RXDELAY_W

Field RXDELAY writer - Sample delay for input serial data on MISO. The value specifies the number of 64 MHz clock cycles (15.625 ns) delay from the the sampling edge of SCK (leading edge for CONFIG.CPHA = 0, trailing edge for CONFIG.CPHA = 1) until the input serial data is sampled. As en example, if RXDELAY = 0 and CONFIG.CPHA = 0, the input serial data is sampled on the rising edge of SCK.