nrf52840_pac::spim0::intenset
Type Alias END_R

pub type END_R = BitReader<END_A>;

Field END reader - Write ‘1’ to enable interrupt for END event

Aliased Type

struct END_R { /* private fields */ }

Implementations

impl END_R

pub fn variant(&self) -> END_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_enabled(&self) -> bool

Checks if the value of the field is ENABLED