nrf52840_pac::uart0::errorsrc

Type Alias PARITY_R

pub type PARITY_R = BitReader<PARITY_A>;

Field PARITY reader - Parity error

Aliased Type

struct PARITY_R { /* private fields */ }

Implementations

impl PARITY_R

pub fn variant(&self) -> PARITY_A

Get enumerated values variant

pub fn is_not_present(&self) -> bool

Checks if the value of the field is NOT_PRESENT

pub fn is_present(&self) -> bool

Checks if the value of the field is PRESENT