nrf52840_pac::uarte0::config

Type Alias STOP_R

pub type STOP_R = BitReader<STOP_A>;

Field STOP reader - Stop bits

Aliased Type

pub struct STOP_R { /* private fields */ }

Implementations

impl STOP_R

pub fn variant(&self) -> STOP_A

Get enumerated values variant

pub fn is_one(&self) -> bool

Checks if the value of the field is ONE

pub fn is_two(&self) -> bool

Checks if the value of the field is TWO