nrf52840_pac::usbd::bmrequesttype

Type Alias RECIPIENT_R

pub type RECIPIENT_R = FieldReader<u8, RECIPIENT_A>;

Field RECIPIENT reader - Data transfer type

Aliased Type

struct RECIPIENT_R { /* private fields */ }

Implementations

impl RECIPIENT_R

pub fn variant(&self) -> Option<RECIPIENT_A>

Get enumerated values variant

pub fn is_device(&self) -> bool

Checks if the value of the field is DEVICE

pub fn is_interface(&self) -> bool

Checks if the value of the field is INTERFACE

pub fn is_endpoint(&self) -> bool

Checks if the value of the field is ENDPOINT

pub fn is_other(&self) -> bool

Checks if the value of the field is OTHER