Type Alias nrf52840_pac::usbd::brequest::BREQUEST_R

pub type BREQUEST_R = FieldReader<u8, BREQUEST_A>;

Field BREQUEST reader - SETUP data, byte 1, bRequest. Values provided for standard requests only, user must implement class and vendor values.

Aliased Type

struct BREQUEST_R { /* private fields */ }

Implementations

impl BREQUEST_R

pub fn variant(&self) -> Option<BREQUEST_A>

Get enumerated values variant

pub fn is_std_get_status(&self) -> bool

Checks if the value of the field is STD_GET_STATUS

pub fn is_std_clear_feature(&self) -> bool

Checks if the value of the field is STD_CLEAR_FEATURE

pub fn is_std_set_feature(&self) -> bool

Checks if the value of the field is STD_SET_FEATURE

pub fn is_std_set_address(&self) -> bool

Checks if the value of the field is STD_SET_ADDRESS

pub fn is_std_get_descriptor(&self) -> bool

Checks if the value of the field is STD_GET_DESCRIPTOR

pub fn is_std_set_descriptor(&self) -> bool

Checks if the value of the field is STD_SET_DESCRIPTOR

pub fn is_std_get_configuration(&self) -> bool

Checks if the value of the field is STD_GET_CONFIGURATION

pub fn is_std_set_configuration(&self) -> bool

Checks if the value of the field is STD_SET_CONFIGURATION

pub fn is_std_get_interface(&self) -> bool

Checks if the value of the field is STD_GET_INTERFACE

pub fn is_std_set_interface(&self) -> bool

Checks if the value of the field is STD_SET_INTERFACE

pub fn is_std_synch_frame(&self) -> bool

Checks if the value of the field is STD_SYNCH_FRAME