nrf52840_pac::usbd
Module dpdmvalue

State D+ and D- lines will be forced into by the DPDMDRIVE task. The DPDMNODRIVE task reverts the control of the lines to MAC IP (no forcing).

Structs

DPDMVALUE_SPEC

State D+ and D- lines will be forced into by the DPDMDRIVE task. The DPDMNODRIVE task reverts the control of the lines to MAC IP (no forcing).

R

Register DPDMVALUE reader

W

Register DPDMVALUE writer

Enums

STATE_A

State D+ and D- lines will be forced into by the DPDMDRIVE task

Type Aliases

STATE_R

Field STATE reader - State D+ and D- lines will be forced into by the DPDMDRIVE task

STATE_W

Field STATE writer - State D+ and D- lines will be forced into by the DPDMDRIVE task