Type Alias nrf52840_pac::usbd::dtoggle::IO_R

pub type IO_R = BitReader<IO_A>;

Field IO reader - Selects IN or OUT endpoint

Aliased Type

struct IO_R { /* private fields */ }

Implementations

impl IO_R

pub fn variant(&self) -> IO_A

Get enumerated values variant

pub fn is_out(&self) -> bool

Checks if the value of the field is OUT

pub fn is_in(&self) -> bool

Checks if the value of the field is IN