nrf52840_pac::usbd::epouten
Type Alias OUT4_R

pub type OUT4_R = BitReader<OUT4_A>;

Field OUT4 reader - Enable OUT endpoint 4

Aliased Type

struct OUT4_R { /* private fields */ }

Implementations

impl OUT4_R

pub fn variant(&self) -> OUT4_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_enable(&self) -> bool

Checks if the value of the field is ENABLE