Module nrf52840_pac::usbd::eventcause

Details on what caused the USBEVENT event

Structs

EVENTCAUSE_SPEC

Details on what caused the USBEVENT event

R

Register EVENTCAUSE reader

W

Register EVENTCAUSE writer

Enums

ISOOUTCRC_A

CRC error was detected on isochronous OUT endpoint 8. Write ‘1’ to clear.

READY_A

USB device is ready for normal operation. Write ‘1’ to clear.

RESUME_A

Signals that a RESUME condition (K state or activity restart) has been detected on USB lines. Write ‘1’ to clear.

SUSPEND_A

Signals that USB lines have been idle long enough for the device to enter suspend. Write ‘1’ to clear.

USBWUALLOWED_A

USB MAC has been woken up and operational. Write ‘1’ to clear.

Type Aliases

ISOOUTCRC_R

Field ISOOUTCRC reader - CRC error was detected on isochronous OUT endpoint 8. Write ‘1’ to clear.

ISOOUTCRC_W

Field ISOOUTCRC writer - CRC error was detected on isochronous OUT endpoint 8. Write ‘1’ to clear.

READY_R

Field READY reader - USB device is ready for normal operation. Write ‘1’ to clear.

READY_W

Field READY writer - USB device is ready for normal operation. Write ‘1’ to clear.

RESUME_R

Field RESUME reader - Signals that a RESUME condition (K state or activity restart) has been detected on USB lines. Write ‘1’ to clear.

RESUME_W

Field RESUME writer - Signals that a RESUME condition (K state or activity restart) has been detected on USB lines. Write ‘1’ to clear.

SUSPEND_R

Field SUSPEND reader - Signals that USB lines have been idle long enough for the device to enter suspend. Write ‘1’ to clear.

SUSPEND_W

Field SUSPEND writer - Signals that USB lines have been idle long enough for the device to enter suspend. Write ‘1’ to clear.

USBWUALLOWED_R

Field USBWUALLOWED reader - USB MAC has been woken up and operational. Write ‘1’ to clear.

USBWUALLOWED_W

Field USBWUALLOWED writer - USB MAC has been woken up and operational. Write ‘1’ to clear.