nrf52840_pac::usbd::isosplit

Type Alias SPLIT_R

pub type SPLIT_R = FieldReader<u16, SPLIT_A>;

Field SPLIT reader - Controls the split of ISO buffers

Aliased Type

struct SPLIT_R { /* private fields */ }

Implementations

impl SPLIT_R

pub fn variant(&self) -> Option<SPLIT_A>

Get enumerated values variant

pub fn is_one_dir(&self) -> bool

Checks if the value of the field is ONE_DIR

pub fn is_half_in(&self) -> bool

Checks if the value of the field is HALF_IN