nrf52840_pac::acl

Module acl

Cluster Unspecified

Modules

addr

Description cluster[n]: Configure the word-aligned start address of region n to protect

perm

Description cluster[n]: Access permissions for region n as defined by start address ACL[n].ADDR and size ACL[n].SIZE

size

Description cluster[n]: Size of region to protect counting from address ACL[n].ADDR. Write ‘0’ as no effect.

unused0

Unspecified

Structs

ACL

Register block

Type Aliases

ADDR

ADDR (rw) register accessor: an alias for Reg<ADDR_SPEC>

PERM

PERM (rw) register accessor: an alias for Reg<PERM_SPEC>

SIZE

SIZE (rw) register accessor: an alias for Reg<SIZE_SPEC>

UNUSED0

UNUSED0 (rw) register accessor: an alias for Reg<UNUSED0_SPEC>