Module nrf52840_pac::acl::acl::perm

Description cluster[n]: Access permissions for region n as defined by start address ACL[n].ADDR and size ACL[n].SIZE

Structs

PERM_SPEC

Description cluster[n]: Access permissions for region n as defined by start address ACL[n].ADDR and size ACL[n].SIZE

R

Register PERM reader

W

Register PERM writer

Enums

READ_A

Configure read permissions for region n. Write ‘0’ has no effect.

WRITE_A

Configure write and erase permissions for region n. Write ‘0’ has no effect.

Type Aliases

READ_R

Field READ reader - Configure read permissions for region n. Write ‘0’ has no effect.

READ_W

Field READ writer - Configure read permissions for region n. Write ‘0’ has no effect.

WRITE_R

Field WRITE reader - Configure write and erase permissions for region n. Write ‘0’ has no effect.

WRITE_W

Field WRITE writer - Configure write and erase permissions for region n. Write ‘0’ has no effect.