nrf52840_pac

Module cc_host_rgf

CRYPTOCELL HOST_RGF interface

Modules

host_cryptokey_sel

AES hardware key select

host_iot_kdr0

This register holds bits 31:0 of K_DR. The value of this register is saved in the CRYPTOCELL AO power domain. Reading from this address returns the K_DR valid status indicating if K_DR is successfully retained.

host_iot_kdr1

This register holds bits 63:32 of K_DR. The value of this register is saved in the CRYPTOCELL AO power domain.

host_iot_kdr2

This register holds bits 95:64 of K_DR. The value of this register is saved in the CRYPTOCELL AO power domain.

host_iot_kdr3

This register holds bits 127:96 of K_DR. The value of this register is saved in the CRYPTOCELL AO power domain.

host_iot_kprtl_lock

This write-once register is the K_PRTL lock register. When this register is set, K_PRTL can not be used and a zeroed key will be used instead. The value of this register is saved in the CRYPTOCELL AO power domain.

host_iot_lcs

Controls lifecycle state (LCS) for CRYPTOCELL subsystem

Structs

RegisterBlock

Register block

Type Aliases

HOST_CRYPTOKEY_SEL

HOST_CRYPTOKEY_SEL (rw) register accessor: an alias for Reg<HOST_CRYPTOKEY_SEL_SPEC>

HOST_IOT_KDR0

HOST_IOT_KDR0 (rw) register accessor: an alias for Reg<HOST_IOT_KDR0_SPEC>

HOST_IOT_KDR1

HOST_IOT_KDR1 (w) register accessor: an alias for Reg<HOST_IOT_KDR1_SPEC>

HOST_IOT_KDR2

HOST_IOT_KDR2 (w) register accessor: an alias for Reg<HOST_IOT_KDR2_SPEC>

HOST_IOT_KDR3

HOST_IOT_KDR3 (w) register accessor: an alias for Reg<HOST_IOT_KDR3_SPEC>

HOST_IOT_KPRTL_LOCK

HOST_IOT_KPRTL_LOCK (rw) register accessor: an alias for Reg<HOST_IOT_KPRTL_LOCK_SPEC>

HOST_IOT_LCS

HOST_IOT_LCS (rw) register accessor: an alias for Reg<HOST_IOT_LCS_SPEC>