nrf52840_pac
Module clock

Clock control

Modules

ctiv

Calibration timer interval

events_ctstarted

Calibration timer has been started and is ready to process new tasks

events_ctstopped

Calibration timer has been stopped and is ready to process new tasks

events_ctto

Calibration timer timeout

events_done

Calibration of LFRC completed

events_hfclkstarted

HFXO crystal oscillator started

events_lfclkstarted

LFCLK started

hfclkrun

Status indicating that HFCLKSTART task has been triggered

hfclkstat

HFCLK status

hfxodebounce

HFXO debounce time. The HFXO is started by triggering the TASKS_HFCLKSTART task.

intenclr

Disable interrupt

intenset

Enable interrupt

lfclkrun

Status indicating that LFCLKSTART task has been triggered

lfclksrc

Clock source for the LFCLK

lfclksrccopy

Copy of LFCLKSRC register, set when LFCLKSTART task was triggered

lfclkstat

LFCLK status

lfrcmode

LFRC mode configuration

tasks_cal

Start calibration of LFRC

tasks_ctstart

Start calibration timer

tasks_ctstop

Stop calibration timer

tasks_hfclkstart

Start HFXO crystal oscillator

tasks_hfclkstop

Stop HFXO crystal oscillator

tasks_lfclkstart

Start LFCLK

tasks_lfclkstop

Stop LFCLK

traceconfig

Clocking options for the trace port debug interface

Structs

RegisterBlock

Register block

Type Aliases

CTIV

CTIV (rw) register accessor: an alias for Reg<CTIV_SPEC>

EVENTS_CTSTARTED

EVENTS_CTSTARTED (rw) register accessor: an alias for Reg<EVENTS_CTSTARTED_SPEC>

EVENTS_CTSTOPPED

EVENTS_CTSTOPPED (rw) register accessor: an alias for Reg<EVENTS_CTSTOPPED_SPEC>

EVENTS_CTTO

EVENTS_CTTO (rw) register accessor: an alias for Reg<EVENTS_CTTO_SPEC>

EVENTS_DONE

EVENTS_DONE (rw) register accessor: an alias for Reg<EVENTS_DONE_SPEC>

EVENTS_HFCLKSTARTED

EVENTS_HFCLKSTARTED (rw) register accessor: an alias for Reg<EVENTS_HFCLKSTARTED_SPEC>

EVENTS_LFCLKSTARTED

EVENTS_LFCLKSTARTED (rw) register accessor: an alias for Reg<EVENTS_LFCLKSTARTED_SPEC>

HFCLKRUN

HFCLKRUN (r) register accessor: an alias for Reg<HFCLKRUN_SPEC>

HFCLKSTAT

HFCLKSTAT (r) register accessor: an alias for Reg<HFCLKSTAT_SPEC>

HFXODEBOUNCE

HFXODEBOUNCE (rw) register accessor: an alias for Reg<HFXODEBOUNCE_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

LFCLKRUN

LFCLKRUN (r) register accessor: an alias for Reg<LFCLKRUN_SPEC>

LFCLKSRC

LFCLKSRC (rw) register accessor: an alias for Reg<LFCLKSRC_SPEC>

LFCLKSRCCOPY

LFCLKSRCCOPY (r) register accessor: an alias for Reg<LFCLKSRCCOPY_SPEC>

LFCLKSTAT

LFCLKSTAT (r) register accessor: an alias for Reg<LFCLKSTAT_SPEC>

LFRCMODE

LFRCMODE (rw) register accessor: an alias for Reg<LFRCMODE_SPEC>

TASKS_CAL

TASKS_CAL (w) register accessor: an alias for Reg<TASKS_CAL_SPEC>

TASKS_CTSTART

TASKS_CTSTART (w) register accessor: an alias for Reg<TASKS_CTSTART_SPEC>

TASKS_CTSTOP

TASKS_CTSTOP (w) register accessor: an alias for Reg<TASKS_CTSTOP_SPEC>

TASKS_HFCLKSTART

TASKS_HFCLKSTART (w) register accessor: an alias for Reg<TASKS_HFCLKSTART_SPEC>

TASKS_HFCLKSTOP

TASKS_HFCLKSTOP (w) register accessor: an alias for Reg<TASKS_HFCLKSTOP_SPEC>

TASKS_LFCLKSTART

TASKS_LFCLKSTART (w) register accessor: an alias for Reg<TASKS_LFCLKSTART_SPEC>

TASKS_LFCLKSTOP

TASKS_LFCLKSTOP (w) register accessor: an alias for Reg<TASKS_LFCLKSTOP_SPEC>

TRACECONFIG

TRACECONFIG (rw) register accessor: an alias for Reg<TRACECONFIG_SPEC>