nrf52840_pac::clock::traceconfig

Type Alias TRACEPORTSPEED_R

pub type TRACEPORTSPEED_R = FieldReader<u8, TRACEPORTSPEED_A>;

Field TRACEPORTSPEED reader - Speed of trace port clock. Note that the TRACECLK pin will output this clock divided by two.

Aliased Type

pub struct TRACEPORTSPEED_R { /* private fields */ }

Implementations

impl TRACEPORTSPEED_R

pub fn variant(&self) -> TRACEPORTSPEED_A

Get enumerated values variant

pub fn is_32mhz(&self) -> bool

Checks if the value of the field is _32MHZ

pub fn is_16mhz(&self) -> bool

Checks if the value of the field is _16MHZ

pub fn is_8mhz(&self) -> bool

Checks if the value of the field is _8MHZ

pub fn is_4mhz(&self) -> bool

Checks if the value of the field is _4MHZ