nrf52840_pac

Module i2s

Inter-IC Sound

Re-exports

pub use config::CONFIG;

pub use rxd::RXD;

pub use txd::TXD;

pub use rxtxd::RXTXD;

pub use psel::PSEL;

Modules

config

Cluster Unspecified

enable

Enable I2S module.

events_rxptrupd

The RXD.PTR register has been copied to internal double-buffers. When the I2S module is started and RX is enabled, this event will be generated for every RXTXD.MAXCNT words that are received on the SDIN pin.

events_stopped

I2S transfer stopped.

events_txptrupd

The TDX.PTR register has been copied to internal double-buffers. When the I2S module is started and TX is enabled, this event will be generated for every RXTXD.MAXCNT words that are sent on the SDOUT pin.

inten

Enable or disable interrupt

intenclr

Disable interrupt

intenset

Enable interrupt

psel

Cluster Unspecified

rxd

Cluster Unspecified

rxtxd

Cluster Unspecified

tasks_start

Starts continuous I2S transfer. Also starts MCK generator when this is enabled.

tasks_stop

Stops I2S transfer. Also stops MCK generator. Triggering this task will cause the {event:STOPPED} event to be generated.

txd

Cluster Unspecified

Structs

RegisterBlock

Register block

Type Aliases

ENABLE

ENABLE (rw) register accessor: an alias for Reg<ENABLE_SPEC>

EVENTS_RXPTRUPD

EVENTS_RXPTRUPD (rw) register accessor: an alias for Reg<EVENTS_RXPTRUPD_SPEC>

EVENTS_STOPPED

EVENTS_STOPPED (rw) register accessor: an alias for Reg<EVENTS_STOPPED_SPEC>

EVENTS_TXPTRUPD

EVENTS_TXPTRUPD (rw) register accessor: an alias for Reg<EVENTS_TXPTRUPD_SPEC>

INTEN

INTEN (rw) register accessor: an alias for Reg<INTEN_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

TASKS_START

TASKS_START (w) register accessor: an alias for Reg<TASKS_START_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>