nrf52840_pac::i2s::config::mckfreq

Type Alias MCKFREQ_R

pub type MCKFREQ_R = FieldReader<u32, MCKFREQ_A>;

Field MCKFREQ reader - Master clock generator frequency.

Aliased Type

struct MCKFREQ_R { /* private fields */ }

Implementations

impl MCKFREQ_R

pub fn variant(&self) -> Option<MCKFREQ_A>

Get enumerated values variant

pub fn is_32mdiv2(&self) -> bool

Checks if the value of the field is _32MDIV2

pub fn is_32mdiv3(&self) -> bool

Checks if the value of the field is _32MDIV3

pub fn is_32mdiv4(&self) -> bool

Checks if the value of the field is _32MDIV4

pub fn is_32mdiv5(&self) -> bool

Checks if the value of the field is _32MDIV5

pub fn is_32mdiv6(&self) -> bool

Checks if the value of the field is _32MDIV6

pub fn is_32mdiv8(&self) -> bool

Checks if the value of the field is _32MDIV8

pub fn is_32mdiv10(&self) -> bool

Checks if the value of the field is _32MDIV10

pub fn is_32mdiv11(&self) -> bool

Checks if the value of the field is _32MDIV11

pub fn is_32mdiv15(&self) -> bool

Checks if the value of the field is _32MDIV15

pub fn is_32mdiv16(&self) -> bool

Checks if the value of the field is _32MDIV16

pub fn is_32mdiv21(&self) -> bool

Checks if the value of the field is _32MDIV21

pub fn is_32mdiv23(&self) -> bool

Checks if the value of the field is _32MDIV23

pub fn is_32mdiv30(&self) -> bool

Checks if the value of the field is _32MDIV30

pub fn is_32mdiv31(&self) -> bool

Checks if the value of the field is _32MDIV31

pub fn is_32mdiv32(&self) -> bool

Checks if the value of the field is _32MDIV32

pub fn is_32mdiv42(&self) -> bool

Checks if the value of the field is _32MDIV42

pub fn is_32mdiv63(&self) -> bool

Checks if the value of the field is _32MDIV63

pub fn is_32mdiv125(&self) -> bool

Checks if the value of the field is _32MDIV125