nrf52840_pac::nfct::txd

Module frameconfig

Configuration of outgoing frames

Structs

FRAMECONFIG_SPEC

Configuration of outgoing frames

R

Register FRAMECONFIG reader

W

Register FRAMECONFIG writer

Enums

CRCMODETX_A

CRC mode for outgoing frames

DISCARDMODE_A

Discarding unused bits at start or end of a frame

PARITY_A

Indicates if parity is added to the frame

SOF_A

Adding SoF or not in TX frames

Type Aliases

CRCMODETX_R

Field CRCMODETX reader - CRC mode for outgoing frames

CRCMODETX_W

Field CRCMODETX writer - CRC mode for outgoing frames

DISCARDMODE_R

Field DISCARDMODE reader - Discarding unused bits at start or end of a frame

DISCARDMODE_W

Field DISCARDMODE writer - Discarding unused bits at start or end of a frame

PARITY_R

Field PARITY reader - Indicates if parity is added to the frame

PARITY_W

Field PARITY writer - Indicates if parity is added to the frame

SOF_R

Field SOF reader - Adding SoF or not in TX frames

SOF_W

Field SOF writer - Adding SoF or not in TX frames