nrf52840_pac::nvmc::eraseuicr

Type Alias ERASEUICR_W

pub type ERASEUICR_W<'a, const O: u8> = BitWriter<'a, u32, ERASEUICR_SPEC, ERASEUICR_A, O>;

Field ERASEUICR writer - Register starting erase of all user information configuration registers. Note that the erase must be enabled using CONFIG.WEN before the UICR can be erased.

Aliased Type

struct ERASEUICR_W<'a, const O: u8> { /* private fields */ }

Implementations

impl<'a, const O: u8> ERASEUICR_W<'a, O>

pub fn no_operation(self) -> &'a mut W

No operation

pub fn erase(self) -> &'a mut W

Start erase of UICR