nrf52840_pac::power::pofcon

Type Alias THRESHOLDVDDH_R

pub type THRESHOLDVDDH_R = FieldReader<u8, THRESHOLDVDDH_A>;

Field THRESHOLDVDDH reader - Power-fail comparator threshold setting for high voltage mode (supply connected to VDDH only). This setting does not apply for normal voltage mode (supply connected to both VDD and VDDH).

Aliased Type

struct THRESHOLDVDDH_R { /* private fields */ }

Implementations

impl THRESHOLDVDDH_R

pub fn variant(&self) -> THRESHOLDVDDH_A

Get enumerated values variant

pub fn is_v27(&self) -> bool

Checks if the value of the field is V27

pub fn is_v28(&self) -> bool

Checks if the value of the field is V28

pub fn is_v29(&self) -> bool

Checks if the value of the field is V29

pub fn is_v30(&self) -> bool

Checks if the value of the field is V30

pub fn is_v31(&self) -> bool

Checks if the value of the field is V31

pub fn is_v32(&self) -> bool

Checks if the value of the field is V32

pub fn is_v33(&self) -> bool

Checks if the value of the field is V33

pub fn is_v34(&self) -> bool

Checks if the value of the field is V34

pub fn is_v35(&self) -> bool

Checks if the value of the field is V35

pub fn is_v36(&self) -> bool

Checks if the value of the field is V36

pub fn is_v37(&self) -> bool

Checks if the value of the field is V37

pub fn is_v38(&self) -> bool

Checks if the value of the field is V38

pub fn is_v39(&self) -> bool

Checks if the value of the field is V39

pub fn is_v40(&self) -> bool

Checks if the value of the field is V40

pub fn is_v41(&self) -> bool

Checks if the value of the field is V41

pub fn is_v42(&self) -> bool

Checks if the value of the field is V42