nrf52840_pac::power

Module resetreas

Reset reason

Structs

R

Register RESETREAS reader

RESETREAS_SPEC

Reset reason

W

Register RESETREAS writer

Enums

DIF_A

Reset due to wake up from System OFF mode when wakeup is triggered from entering into debug interface mode

DOG_A

Reset from watchdog detected

LOCKUP_A

Reset from CPU lock-up detected

LPCOMP_A

Reset due to wake up from System OFF mode when wakeup is triggered from ANADETECT signal from LPCOMP

NFC_A

Reset due to wake up from System OFF mode by NFC field detect

OFF_A

Reset due to wake up from System OFF mode when wakeup is triggered from DETECT signal from GPIO

RESETPIN_A

Reset from pin-reset detected

SREQ_A

Reset from soft reset detected

VBUS_A

Reset due to wake up from System OFF mode by VBUS rising into valid range

Type Aliases

DIF_R

Field DIF reader - Reset due to wake up from System OFF mode when wakeup is triggered from entering into debug interface mode

DIF_W

Field DIF writer - Reset due to wake up from System OFF mode when wakeup is triggered from entering into debug interface mode

DOG_R

Field DOG reader - Reset from watchdog detected

DOG_W

Field DOG writer - Reset from watchdog detected

LOCKUP_R

Field LOCKUP reader - Reset from CPU lock-up detected

LOCKUP_W

Field LOCKUP writer - Reset from CPU lock-up detected

LPCOMP_R

Field LPCOMP reader - Reset due to wake up from System OFF mode when wakeup is triggered from ANADETECT signal from LPCOMP

LPCOMP_W

Field LPCOMP writer - Reset due to wake up from System OFF mode when wakeup is triggered from ANADETECT signal from LPCOMP

NFC_R

Field NFC reader - Reset due to wake up from System OFF mode by NFC field detect

NFC_W

Field NFC writer - Reset due to wake up from System OFF mode by NFC field detect

OFF_R

Field OFF reader - Reset due to wake up from System OFF mode when wakeup is triggered from DETECT signal from GPIO

OFF_W

Field OFF writer - Reset due to wake up from System OFF mode when wakeup is triggered from DETECT signal from GPIO

RESETPIN_R

Field RESETPIN reader - Reset from pin-reset detected

RESETPIN_W

Field RESETPIN writer - Reset from pin-reset detected

SREQ_R

Field SREQ reader - Reset from soft reset detected

SREQ_W

Field SREQ writer - Reset from soft reset detected

VBUS_R

Field VBUS reader - Reset due to wake up from System OFF mode by VBUS rising into valid range

VBUS_W

Field VBUS writer - Reset due to wake up from System OFF mode by VBUS rising into valid range