nrf52840_pac

Module pwm0

Pulse width modulation unit 0

Re-exports

pub use seq::SEQ;

pub use psel::PSEL;

Modules

countertop

Value up to which the pulse generator counter counts

decoder

Configuration of the decoder

enable

PWM module enable register

events_loopsdone

Concatenated sequences have been played the amount of times defined in LOOP.CNT

events_pwmperiodend

Emitted at the end of each PWM period

events_seqend

Description collection[n]: Emitted at end of every sequence n, when last value from RAM has been applied to wave counter

events_seqstarted

Description collection[n]: First PWM period started on sequence n

events_stopped

Response to STOP task, emitted when PWM pulses are no longer generated

inten

Enable or disable interrupt

intenclr

Disable interrupt

intenset

Enable interrupt

loop_

Number of playbacks of a loop

mode

Selects operating mode of the wave counter

prescaler

Configuration for PWM_CLK

psel

Cluster Unspecified

seq

Cluster Unspecified

shorts

Shortcut register

tasks_nextstep

Steps by one value in the current sequence on all enabled channels if DECODER.MODE=NextStep. Does not cause PWM generation to start if not running.

tasks_seqstart

Description collection[n]: Loads the first PWM value on all enabled channels from sequence n, and starts playing that sequence at the rate defined in SEQ[n]REFRESH and/or DECODER.MODE. Causes PWM generation to start if not running.

tasks_stop

Stops PWM pulse generation on all channels at the end of current PWM period, and stops sequence playback

Structs

RegisterBlock

Register block

Type Aliases

COUNTERTOP

COUNTERTOP (rw) register accessor: an alias for Reg<COUNTERTOP_SPEC>

DECODER

DECODER (rw) register accessor: an alias for Reg<DECODER_SPEC>

ENABLE

ENABLE (rw) register accessor: an alias for Reg<ENABLE_SPEC>

EVENTS_LOOPSDONE

EVENTS_LOOPSDONE (rw) register accessor: an alias for Reg<EVENTS_LOOPSDONE_SPEC>

EVENTS_PWMPERIODEND

EVENTS_PWMPERIODEND (rw) register accessor: an alias for Reg<EVENTS_PWMPERIODEND_SPEC>

EVENTS_SEQEND

EVENTS_SEQEND (rw) register accessor: an alias for Reg<EVENTS_SEQEND_SPEC>

EVENTS_SEQSTARTED

EVENTS_SEQSTARTED (rw) register accessor: an alias for Reg<EVENTS_SEQSTARTED_SPEC>

EVENTS_STOPPED

EVENTS_STOPPED (rw) register accessor: an alias for Reg<EVENTS_STOPPED_SPEC>

INTEN

INTEN (rw) register accessor: an alias for Reg<INTEN_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

LOOP

LOOP (rw) register accessor: an alias for Reg<LOOP_SPEC>

MODE

MODE (rw) register accessor: an alias for Reg<MODE_SPEC>

PRESCALER

PRESCALER (rw) register accessor: an alias for Reg<PRESCALER_SPEC>

SHORTS

SHORTS (rw) register accessor: an alias for Reg<SHORTS_SPEC>

TASKS_NEXTSTEP

TASKS_NEXTSTEP (w) register accessor: an alias for Reg<TASKS_NEXTSTEP_SPEC>

TASKS_SEQSTART

TASKS_SEQSTART (w) register accessor: an alias for Reg<TASKS_SEQSTART_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>