nrf52840_pac::qspi

Module cinstrconf

Custom instruction configuration register.

Structs

CINSTRCONF_SPEC

Custom instruction configuration register.

R

Register CINSTRCONF reader

W

Register CINSTRCONF writer

Enums

LENGTH_A

Length of custom instruction in number of bytes.

LFEN_A

Enable long frame mode. When enabled, a custom instruction transaction has to be ended by writing the LFSTOP field.

LFSTOP_A

Stop (finalize) long frame transaction

WIPWAIT_A

Wait for write complete before sending command.

WREN_A

Send WREN (write enable opcode 0x06) before instruction.

Type Aliases

LENGTH_R

Field LENGTH reader - Length of custom instruction in number of bytes.

LENGTH_W

Field LENGTH writer - Length of custom instruction in number of bytes.

LFEN_R

Field LFEN reader - Enable long frame mode. When enabled, a custom instruction transaction has to be ended by writing the LFSTOP field.

LFEN_W

Field LFEN writer - Enable long frame mode. When enabled, a custom instruction transaction has to be ended by writing the LFSTOP field.

LFSTOP_R

Field LFSTOP reader - Stop (finalize) long frame transaction

LFSTOP_W

Field LFSTOP writer - Stop (finalize) long frame transaction

LIO2_R

Field LIO2 reader - Level of the IO2 pin (if connected) during transmission of custom instruction.

LIO2_W

Field LIO2 writer - Level of the IO2 pin (if connected) during transmission of custom instruction.

LIO3_R

Field LIO3 reader - Level of the IO3 pin (if connected) during transmission of custom instruction.

LIO3_W

Field LIO3 writer - Level of the IO3 pin (if connected) during transmission of custom instruction.

OPCODE_R

Field OPCODE reader - Opcode of Custom instruction.

OPCODE_W

Field OPCODE writer - Opcode of Custom instruction.

WIPWAIT_R

Field WIPWAIT reader - Wait for write complete before sending command.

WIPWAIT_W

Field WIPWAIT writer - Wait for write complete before sending command.

WREN_R

Field WREN reader - Send WREN (write enable opcode 0x06) before instruction.

WREN_W

Field WREN writer - Send WREN (write enable opcode 0x06) before instruction.