nrf52840_pac

Module radio

2.4 GHz radio

Modules

base0

Base address 0

base1

Base address 1

bcc

Bit counter compare

ccactrl

IEEE 802.15.4 clear channel assessment control

crccnf

CRC configuration

crcinit

CRC initial value

crcpoly

CRC polynomial

crcstatus

CRC status

dab

Description collection[n]: Device address base segment n

dacnf

Device address match configuration

dai

Device address match index

dap

Description collection[n]: Device address prefix n

datawhiteiv

Data whitening initial value

edcnt

IEEE 802.15.4 energy detect loop count

edsample

IEEE 802.15.4 energy detect level

events_address

Address sent or received

events_bcmatch

Bit counter reached bit count value

events_ccabusy

Wireless medium busy - do not send

events_ccaidle

Wireless medium in idle - clear to send

events_ccastopped

The CCA has stopped

events_crcerror

Packet received with CRC error

events_crcok

Packet received with CRC ok

events_devmatch

A device address match occurred on the last received packet

events_devmiss

No device address match occurred on the last received packet

events_disabled

RADIO has been disabled

events_edend

Sampling of energy detection complete. A new ED sample is ready for readout from the RADIO.EDSAMPLE register.

events_edstopped

The sampling of energy detection has stopped

events_end

Packet sent or received

events_framestart

IEEE 802.15.4 length field received

events_mhrmatch

MAC header match found

events_payload

Packet payload sent or received

events_phyend

Generated in Ble_LR125Kbit, Ble_LR500Kbit and BleIeee802154_250Kbit modes when last bit is sent on air.

events_rateboost

Ble_LR CI field received, receive mode is changed from Ble_LR125Kbit to Ble_LR500Kbit.

events_ready

RADIO has ramped up and is ready to be started

events_rssiend

Sampling of receive signal strength complete

events_rxready

RADIO has ramped up and is ready to be started RX path

events_txready

RADIO has ramped up and is ready to be started TX path

frequency

Frequency

intenclr

Disable interrupt

intenset

Enable interrupt

mhrmatchconf

Search pattern configuration

mhrmatchmas

Pattern mask

mode

Data rate and modulation

modecnf0

Radio mode configuration register 0

packetptr

Packet pointer

pcnf0

Packet configuration register 0

pcnf1

Packet configuration register 1

pdustat

Payload status

power

Peripheral power control

prefix0

Prefixes bytes for logical addresses 0-3

prefix1

Prefixes bytes for logical addresses 4-7

rssisample

RSSI sample

rxaddresses

Receive address select

rxcrc

CRC field of previously received packet

rxmatch

Received address

sfd

IEEE 802.15.4 start of frame delimiter

shorts

Shortcut register

state

Current radio state

tasks_bcstart

Start the bit counter

tasks_bcstop

Stop the bit counter

tasks_ccastart

Start the clear channel assessment used in IEEE 802.15.4 mode

tasks_ccastop

Stop the clear channel assessment

tasks_disable

Disable RADIO

tasks_edstart

Start the energy detect measurement used in IEEE 802.15.4 mode

tasks_edstop

Stop the energy detect measurement

tasks_rssistart

Start the RSSI and take one single sample of the receive signal strength

tasks_rssistop

Stop the RSSI measurement

tasks_rxen

Enable RADIO in RX mode

tasks_start

Start RADIO

tasks_stop

Stop RADIO

tasks_txen

Enable RADIO in TX mode

tifs

Interframe spacing in us

txaddress

Transmit address select

txpower

Output power

Structs

RegisterBlock

Register block

Type Aliases

BASE0

BASE0 (rw) register accessor: an alias for Reg<BASE0_SPEC>

BASE1

BASE1 (rw) register accessor: an alias for Reg<BASE1_SPEC>

BCC

BCC (rw) register accessor: an alias for Reg<BCC_SPEC>

CCACTRL

CCACTRL (rw) register accessor: an alias for Reg<CCACTRL_SPEC>

CRCCNF

CRCCNF (rw) register accessor: an alias for Reg<CRCCNF_SPEC>

CRCINIT

CRCINIT (rw) register accessor: an alias for Reg<CRCINIT_SPEC>

CRCPOLY

CRCPOLY (rw) register accessor: an alias for Reg<CRCPOLY_SPEC>

CRCSTATUS

CRCSTATUS (r) register accessor: an alias for Reg<CRCSTATUS_SPEC>

DAB

DAB (rw) register accessor: an alias for Reg<DAB_SPEC>

DACNF

DACNF (rw) register accessor: an alias for Reg<DACNF_SPEC>

DAI

DAI (r) register accessor: an alias for Reg<DAI_SPEC>

DAP

DAP (rw) register accessor: an alias for Reg<DAP_SPEC>

DATAWHITEIV

DATAWHITEIV (rw) register accessor: an alias for Reg<DATAWHITEIV_SPEC>

EDCNT

EDCNT (rw) register accessor: an alias for Reg<EDCNT_SPEC>

EDSAMPLE

EDSAMPLE (rw) register accessor: an alias for Reg<EDSAMPLE_SPEC>

EVENTS_ADDRESS

EVENTS_ADDRESS (rw) register accessor: an alias for Reg<EVENTS_ADDRESS_SPEC>

EVENTS_BCMATCH

EVENTS_BCMATCH (rw) register accessor: an alias for Reg<EVENTS_BCMATCH_SPEC>

EVENTS_CCABUSY

EVENTS_CCABUSY (rw) register accessor: an alias for Reg<EVENTS_CCABUSY_SPEC>

EVENTS_CCAIDLE

EVENTS_CCAIDLE (rw) register accessor: an alias for Reg<EVENTS_CCAIDLE_SPEC>

EVENTS_CCASTOPPED

EVENTS_CCASTOPPED (rw) register accessor: an alias for Reg<EVENTS_CCASTOPPED_SPEC>

EVENTS_CRCERROR

EVENTS_CRCERROR (rw) register accessor: an alias for Reg<EVENTS_CRCERROR_SPEC>

EVENTS_CRCOK

EVENTS_CRCOK (rw) register accessor: an alias for Reg<EVENTS_CRCOK_SPEC>

EVENTS_DEVMATCH

EVENTS_DEVMATCH (rw) register accessor: an alias for Reg<EVENTS_DEVMATCH_SPEC>

EVENTS_DEVMISS

EVENTS_DEVMISS (rw) register accessor: an alias for Reg<EVENTS_DEVMISS_SPEC>

EVENTS_DISABLED

EVENTS_DISABLED (rw) register accessor: an alias for Reg<EVENTS_DISABLED_SPEC>

EVENTS_EDEND

EVENTS_EDEND (rw) register accessor: an alias for Reg<EVENTS_EDEND_SPEC>

EVENTS_EDSTOPPED

EVENTS_EDSTOPPED (rw) register accessor: an alias for Reg<EVENTS_EDSTOPPED_SPEC>

EVENTS_END

EVENTS_END (rw) register accessor: an alias for Reg<EVENTS_END_SPEC>

EVENTS_FRAMESTART

EVENTS_FRAMESTART (rw) register accessor: an alias for Reg<EVENTS_FRAMESTART_SPEC>

EVENTS_MHRMATCH

EVENTS_MHRMATCH (rw) register accessor: an alias for Reg<EVENTS_MHRMATCH_SPEC>

EVENTS_PAYLOAD

EVENTS_PAYLOAD (rw) register accessor: an alias for Reg<EVENTS_PAYLOAD_SPEC>

EVENTS_PHYEND

EVENTS_PHYEND (rw) register accessor: an alias for Reg<EVENTS_PHYEND_SPEC>

EVENTS_RATEBOOST

EVENTS_RATEBOOST (rw) register accessor: an alias for Reg<EVENTS_RATEBOOST_SPEC>

EVENTS_READY

EVENTS_READY (rw) register accessor: an alias for Reg<EVENTS_READY_SPEC>

EVENTS_RSSIEND

EVENTS_RSSIEND (rw) register accessor: an alias for Reg<EVENTS_RSSIEND_SPEC>

EVENTS_RXREADY

EVENTS_RXREADY (rw) register accessor: an alias for Reg<EVENTS_RXREADY_SPEC>

EVENTS_TXREADY

EVENTS_TXREADY (rw) register accessor: an alias for Reg<EVENTS_TXREADY_SPEC>

FREQUENCY

FREQUENCY (rw) register accessor: an alias for Reg<FREQUENCY_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

MHRMATCHCONF

MHRMATCHCONF (rw) register accessor: an alias for Reg<MHRMATCHCONF_SPEC>

MHRMATCHMAS

MHRMATCHMAS (rw) register accessor: an alias for Reg<MHRMATCHMAS_SPEC>

MODE

MODE (rw) register accessor: an alias for Reg<MODE_SPEC>

MODECNF0

MODECNF0 (rw) register accessor: an alias for Reg<MODECNF0_SPEC>

PACKETPTR

PACKETPTR (rw) register accessor: an alias for Reg<PACKETPTR_SPEC>

PCNF0

PCNF0 (rw) register accessor: an alias for Reg<PCNF0_SPEC>

PCNF1

PCNF1 (rw) register accessor: an alias for Reg<PCNF1_SPEC>

PDUSTAT

PDUSTAT (r) register accessor: an alias for Reg<PDUSTAT_SPEC>

POWER

POWER (rw) register accessor: an alias for Reg<POWER_SPEC>

PREFIX0

PREFIX0 (rw) register accessor: an alias for Reg<PREFIX0_SPEC>

PREFIX1

PREFIX1 (rw) register accessor: an alias for Reg<PREFIX1_SPEC>

RSSISAMPLE

RSSISAMPLE (r) register accessor: an alias for Reg<RSSISAMPLE_SPEC>

RXADDRESSES

RXADDRESSES (rw) register accessor: an alias for Reg<RXADDRESSES_SPEC>

RXCRC

RXCRC (r) register accessor: an alias for Reg<RXCRC_SPEC>

RXMATCH

RXMATCH (r) register accessor: an alias for Reg<RXMATCH_SPEC>

SFD

SFD (rw) register accessor: an alias for Reg<SFD_SPEC>

SHORTS

SHORTS (rw) register accessor: an alias for Reg<SHORTS_SPEC>

STATE

STATE (r) register accessor: an alias for Reg<STATE_SPEC>

TASKS_BCSTART

TASKS_BCSTART (w) register accessor: an alias for Reg<TASKS_BCSTART_SPEC>

TASKS_BCSTOP

TASKS_BCSTOP (w) register accessor: an alias for Reg<TASKS_BCSTOP_SPEC>

TASKS_CCASTART

TASKS_CCASTART (w) register accessor: an alias for Reg<TASKS_CCASTART_SPEC>

TASKS_CCASTOP

TASKS_CCASTOP (w) register accessor: an alias for Reg<TASKS_CCASTOP_SPEC>

TASKS_DISABLE

TASKS_DISABLE (w) register accessor: an alias for Reg<TASKS_DISABLE_SPEC>

TASKS_EDSTART

TASKS_EDSTART (w) register accessor: an alias for Reg<TASKS_EDSTART_SPEC>

TASKS_EDSTOP

TASKS_EDSTOP (w) register accessor: an alias for Reg<TASKS_EDSTOP_SPEC>

TASKS_RSSISTART

TASKS_RSSISTART (w) register accessor: an alias for Reg<TASKS_RSSISTART_SPEC>

TASKS_RSSISTOP

TASKS_RSSISTOP (w) register accessor: an alias for Reg<TASKS_RSSISTOP_SPEC>

TASKS_RXEN

TASKS_RXEN (w) register accessor: an alias for Reg<TASKS_RXEN_SPEC>

TASKS_START

TASKS_START (w) register accessor: an alias for Reg<TASKS_START_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>

TASKS_TXEN

TASKS_TXEN (w) register accessor: an alias for Reg<TASKS_TXEN_SPEC>

TIFS

TIFS (rw) register accessor: an alias for Reg<TIFS_SPEC>

TXADDRESS

TXADDRESS (rw) register accessor: an alias for Reg<TXADDRESS_SPEC>

TXPOWER

TXPOWER (rw) register accessor: an alias for Reg<TXPOWER_SPEC>