Module nrf52840_pac::radio::dacnf

Device address match configuration

Structs

DACNF_SPEC

Device address match configuration

R

Register DACNF reader

W

Register DACNF writer

Enums

ENA0_A

Enable or disable device address matching using device address 0

ENA1_A

Enable or disable device address matching using device address 1

ENA2_A

Enable or disable device address matching using device address 2

ENA3_A

Enable or disable device address matching using device address 3

ENA4_A

Enable or disable device address matching using device address 4

ENA5_A

Enable or disable device address matching using device address 5

ENA6_A

Enable or disable device address matching using device address 6

ENA7_A

Enable or disable device address matching using device address 7

Type Aliases

ENA0_R

Field ENA0 reader - Enable or disable device address matching using device address 0

ENA0_W

Field ENA0 writer - Enable or disable device address matching using device address 0

ENA1_R

Field ENA1 reader - Enable or disable device address matching using device address 1

ENA1_W

Field ENA1 writer - Enable or disable device address matching using device address 1

ENA2_R

Field ENA2 reader - Enable or disable device address matching using device address 2

ENA2_W

Field ENA2 writer - Enable or disable device address matching using device address 2

ENA3_R

Field ENA3 reader - Enable or disable device address matching using device address 3

ENA3_W

Field ENA3 writer - Enable or disable device address matching using device address 3

ENA4_R

Field ENA4 reader - Enable or disable device address matching using device address 4

ENA4_W

Field ENA4 writer - Enable or disable device address matching using device address 4

ENA5_R

Field ENA5 reader - Enable or disable device address matching using device address 5

ENA5_W

Field ENA5 writer - Enable or disable device address matching using device address 5

ENA6_R

Field ENA6 reader - Enable or disable device address matching using device address 6

ENA6_W

Field ENA6 writer - Enable or disable device address matching using device address 6

ENA7_R

Field ENA7 reader - Enable or disable device address matching using device address 7

ENA7_W

Field ENA7 writer - Enable or disable device address matching using device address 7

TXADD0_R

Field TXADD0 reader - TxAdd for device address 0

TXADD0_W

Field TXADD0 writer - TxAdd for device address 0

TXADD1_R

Field TXADD1 reader - TxAdd for device address 1

TXADD1_W

Field TXADD1 writer - TxAdd for device address 1

TXADD2_R

Field TXADD2 reader - TxAdd for device address 2

TXADD2_W

Field TXADD2 writer - TxAdd for device address 2

TXADD3_R

Field TXADD3 reader - TxAdd for device address 3

TXADD3_W

Field TXADD3 writer - TxAdd for device address 3

TXADD4_R

Field TXADD4 reader - TxAdd for device address 4

TXADD4_W

Field TXADD4 writer - TxAdd for device address 4

TXADD5_R

Field TXADD5 reader - TxAdd for device address 5

TXADD5_W

Field TXADD5 writer - TxAdd for device address 5

TXADD6_R

Field TXADD6 reader - TxAdd for device address 6

TXADD6_W

Field TXADD6 writer - TxAdd for device address 6

TXADD7_R

Field TXADD7 reader - TxAdd for device address 7

TXADD7_W

Field TXADD7 writer - TxAdd for device address 7