nrf52840_pac

Module spim0

Serial Peripheral Interface Master with EasyDMA 0

Re-exports

pub use psel::PSEL;

pub use rxd::RXD;

pub use txd::TXD;

pub use iftiming::IFTIMING;

Modules

config

Configuration register

csnpol

Polarity of CSN output

dcxcnt

DCX configuration

enable

Enable SPIM

events_end

End of RXD buffer and TXD buffer reached

events_endrx

End of RXD buffer reached

events_endtx

End of TXD buffer reached

events_started

Transaction started

events_stopped

SPI transaction has stopped

frequency

SPI frequency. Accuracy depends on the HFCLK source selected.

iftiming

Cluster Unspecified

intenclr

Disable interrupt

intenset

Enable interrupt

orc

Byte transmitted after TXD.MAXCNT bytes have been transmitted in the case when RXD.MAXCNT is greater than TXD.MAXCNT

psel

Cluster Unspecified

pseldcx

Pin select for DCX signal

rxd

Cluster RXD EasyDMA channel

shorts

Shortcut register

stallstat

Stall status for EasyDMA RAM accesses. The fields in this register is set to STALL by hardware whenever a stall occurres and can be cleared (set to NOSTALL) by the CPU.

tasks_resume

Resume SPI transaction

tasks_start

Start SPI transaction

tasks_stop

Stop SPI transaction

tasks_suspend

Suspend SPI transaction

txd

Cluster TXD EasyDMA channel

Structs

RegisterBlock

Register block

Type Aliases

CONFIG

CONFIG (rw) register accessor: an alias for Reg<CONFIG_SPEC>

CSNPOL

CSNPOL (rw) register accessor: an alias for Reg<CSNPOL_SPEC>

DCXCNT

DCXCNT (rw) register accessor: an alias for Reg<DCXCNT_SPEC>

ENABLE

ENABLE (rw) register accessor: an alias for Reg<ENABLE_SPEC>

EVENTS_END

EVENTS_END (rw) register accessor: an alias for Reg<EVENTS_END_SPEC>

EVENTS_ENDRX

EVENTS_ENDRX (rw) register accessor: an alias for Reg<EVENTS_ENDRX_SPEC>

EVENTS_ENDTX

EVENTS_ENDTX (rw) register accessor: an alias for Reg<EVENTS_ENDTX_SPEC>

EVENTS_STARTED

EVENTS_STARTED (rw) register accessor: an alias for Reg<EVENTS_STARTED_SPEC>

EVENTS_STOPPED

EVENTS_STOPPED (rw) register accessor: an alias for Reg<EVENTS_STOPPED_SPEC>

FREQUENCY

FREQUENCY (rw) register accessor: an alias for Reg<FREQUENCY_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

ORC

ORC (rw) register accessor: an alias for Reg<ORC_SPEC>

PSELDCX

PSELDCX (rw) register accessor: an alias for Reg<PSELDCX_SPEC>

SHORTS

SHORTS (rw) register accessor: an alias for Reg<SHORTS_SPEC>

STALLSTAT

STALLSTAT (rw) register accessor: an alias for Reg<STALLSTAT_SPEC>

TASKS_RESUME

TASKS_RESUME (w) register accessor: an alias for Reg<TASKS_RESUME_SPEC>

TASKS_START

TASKS_START (w) register accessor: an alias for Reg<TASKS_START_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>

TASKS_SUSPEND

TASKS_SUSPEND (w) register accessor: an alias for Reg<TASKS_SUSPEND_SPEC>