Type Alias nrf52840_pac::uarte0::baudrate::BAUDRATE_R

pub type BAUDRATE_R = FieldReader<u32, BAUDRATE_A>;

Field BAUDRATE reader - Baud rate

Aliased Type

struct BAUDRATE_R { /* private fields */ }

Implementations

impl BAUDRATE_R

pub fn variant(&self) -> Option<BAUDRATE_A>

Get enumerated values variant

pub fn is_baud1200(&self) -> bool

Checks if the value of the field is BAUD1200

pub fn is_baud2400(&self) -> bool

Checks if the value of the field is BAUD2400

pub fn is_baud4800(&self) -> bool

Checks if the value of the field is BAUD4800

pub fn is_baud9600(&self) -> bool

Checks if the value of the field is BAUD9600

pub fn is_baud14400(&self) -> bool

Checks if the value of the field is BAUD14400

pub fn is_baud19200(&self) -> bool

Checks if the value of the field is BAUD19200

pub fn is_baud28800(&self) -> bool

Checks if the value of the field is BAUD28800

pub fn is_baud31250(&self) -> bool

Checks if the value of the field is BAUD31250

pub fn is_baud38400(&self) -> bool

Checks if the value of the field is BAUD38400

pub fn is_baud56000(&self) -> bool

Checks if the value of the field is BAUD56000

pub fn is_baud57600(&self) -> bool

Checks if the value of the field is BAUD57600

pub fn is_baud76800(&self) -> bool

Checks if the value of the field is BAUD76800

pub fn is_baud115200(&self) -> bool

Checks if the value of the field is BAUD115200

pub fn is_baud230400(&self) -> bool

Checks if the value of the field is BAUD230400

pub fn is_baud250000(&self) -> bool

Checks if the value of the field is BAUD250000

pub fn is_baud460800(&self) -> bool

Checks if the value of the field is BAUD460800

pub fn is_baud921600(&self) -> bool

Checks if the value of the field is BAUD921600

pub fn is_baud1m(&self) -> bool

Checks if the value of the field is BAUD1M