nrf52840_pac::uicr::regout0

Type Alias VOUT_R

pub type VOUT_R = FieldReader<u8, VOUT_A>;

Field VOUT reader - Output voltage from of REG0 regulator stage. The maximum output voltage from this stage is given as VDDH - VEXDIF.

Aliased Type

struct VOUT_R { /* private fields */ }

Implementations

impl VOUT_R

pub fn variant(&self) -> Option<VOUT_A>

Get enumerated values variant

pub fn is_1v8(&self) -> bool

Checks if the value of the field is _1V8

pub fn is_2v1(&self) -> bool

Checks if the value of the field is _2V1

pub fn is_2v4(&self) -> bool

Checks if the value of the field is _2V4

pub fn is_2v7(&self) -> bool

Checks if the value of the field is _2V7

pub fn is_3v0(&self) -> bool

Checks if the value of the field is _3V0

pub fn is_3v3(&self) -> bool

Checks if the value of the field is _3V3

pub fn is_default(&self) -> bool

Checks if the value of the field is DEFAULT